### 2025-07-21 #8630 @editorialbot
[PRE REVIEW]: rojak: A Python library and tool for aviation turbulence diagnostics
**Submitting author:** @wong-hl (<a href="http://orcid.org/0009-0006-2314-2660">Hui Ling Wong</a>)
**Repository:** https://github.com/ImperialCollegeLondon/rojak
**Branch with paper.md** (empty if default branch): joss-paper
**Version:** v0.1.1
