### 2025-07-22 c26607a
Added JOSS status badge
