### 2025-07-22 858f424
feat: Put in the plumbing to compute ROC curves from the CLI
