+++ src/rojak/orchestrator/turbulence.py
@@ -28,6 +28,7 @@ from rojak.datalib.madis.amdar import AcarsAmdarRepository
 from rojak.datalib.ukmo.amdar import UkmoAmdarRepository
 from rojak.orchestrator.configuration import (
     AmdarDataSource,
+    DiagnosticValidationCondition,
     TurbulenceCalibrationPhaseOption,
     TurbulenceEvaluationConfig,
     TurbulenceEvaluationPhaseOption,
@@ -38,6 +39,7 @@ from rojak.plot.turbulence_plotter import (
     create_diagnostic_correlation_plot,
     create_multi_region_correlation_plot,
     create_multi_turbulence_diagnotics_probability_plot,
+    plot_roc_curve,
 )
 from rojak.turbulence.analysis import (
     CorrelationBetweenDiagnostics,
@@ -45,12 +47,17 @@ from rojak.turbulence.analysis import (
     LatitudinalCorrelationBetweenDiagnostics,
 )
 from rojak.turbulence.diagnostic import CalibrationDiagnosticSuite, DiagnosticFactory, EvaluationDiagnosticSuite
-from rojak.turbulence.verification import DiagnosticsAmdarDataHarmoniser, DiagnosticsAmdarHarmonisationStrategyOptions
+from rojak.turbulence.verification import (
+    DiagnosticsAmdarDataHarmoniser,
+    DiagnosticsAmdarHarmonisationStrategyOptions,
+    DiagnosticsAmdarVerification,
+)
 from rojak.utilities.types import DistributionParameters, Limits
 
 if TYPE_CHECKING:
     from pathlib import Path
 
+    import dask.array as da
     import dask.dataframe as dd
 
     from rojak.core.data import AmdarDataRepository, CATData
@@ -62,6 +69,9 @@ if TYPE_CHECKING:
         TurbulenceCalibrationPhases,
         TurbulenceConfig,
         TurbulenceDiagnostics,
+    )
+    from rojak.turbulence.verification import (
+        RocVerificationResult,
     )
     from rojak.utilities.types import DiagnosticName
 
@@ -415,14 +425,24 @@ class DiagnosticsAmdarLauncher:
     _time_window: "Limits[datetime]"
     _output_filepath: "Path"
     _save_output: bool
+    _validation_conditions: list["DiagnosticValidationCondition"]
 
     def __init__(self, data_config: "DataConfig", output_dir: "Path", run_name: "RunName") -> None:
         assert data_config.amdar_config is not None
         self._data_source = data_config.amdar_config.data_source
         self._path_to_files = str(data_config.amdar_config.data_dir.resolve() / data_config.amdar_config.glob_pattern)
         self._spatial_domain = data_config.spatial_domain
-        self._strategies = data_config.amdar_config.harmonisation_strategies
+        self._strategies = (
+            data_config.amdar_config.harmonisation_strategies
+            if data_config.amdar_config.harmonisation_strategies is not None
+            else []
+        )
         self._time_window = data_config.amdar_config.time_window
+        self._validation_conditions = (
+            []
+            if data_config.amdar_config.diagnostic_validation is None
+            else data_config.amdar_config.diagnostic_validation.validation_conditions
+        )
 
         self._save_output = data_config.amdar_config.save_harmonised_data
         base_dir = output_dir / run_name / "data_harmonisation"
@@ -443,7 +463,7 @@ class DiagnosticsAmdarLauncher:
             case _ as unreachable:
                 assert_never(unreachable)
 
-    def launch(self, diagnostic_suite: EvaluationDiagnosticSuite) -> "dd.DataFrame":
+    def launch(self, diagnostic_suite: EvaluationDiagnosticSuite) -> None:
         if self._spatial_domain.grid_size is None:
[…]
