### 2025-07-22 e0ef71b
test: Added integration test which runs the calibration stage and checks the generated files. This test highlighted the bug that was fixed in the previous commit
