+++ tests/integration/test_turbulence_launcher.py
@@ -1,0 +1,116 @@
+import random
+from collections.abc import Callable
+from pathlib import Path
+from typing import TYPE_CHECKING, NamedTuple
+
+import pytest
+
+from rojak.orchestrator.configuration import (
+    Context as ConfigContext,
+)
+from rojak.orchestrator.configuration import (
+    DataConfig,
+    SpatialDomain,
+    TurbulenceCalibrationConfig,
+    TurbulenceCalibrationPhaseOption,
+    TurbulenceCalibrationPhases,
+    TurbulenceConfig,
+    TurbulenceDiagnostics,
+    TurbulencePhases,
+    TurbulenceThresholds,
+)
+from rojak.orchestrator.turbulence import DISTRIBUTION_PARAMS_TYPE_ADAPTER, THRESHOLDS_TYPE_ADAPTER, TurbulenceLauncher
+
+if TYPE_CHECKING:
+    from rojak.turbulence.analysis import HistogramData
+
+
+@pytest.fixture
+def create_calibration_only_config(tmp_path_factory) -> Callable:
+    def _calibration_only_config(num_diagnostics: int) -> ConfigContext:
+        plots_dir: Path = tmp_path_factory.mktemp("plots")
+        output_dir: Path = tmp_path_factory.mktemp("output")
+        diagnostics: list[TurbulenceDiagnostics] = random.sample(list(TurbulenceDiagnostics), k=num_diagnostics)
+        return ConfigContext(
+            name="calibration_only",
+            image_format="png",
+            output_dir=output_dir,
+            plots_dir=plots_dir,
+            turbulence_config=TurbulenceConfig(
+                chunks={"pressure_level": 3, "latitude": 721, "longitude": 1440},
+                diagnostics=diagnostics,
+                phases=TurbulencePhases(
+                    calibration_phases=TurbulenceCalibrationPhases(
+                        phases=[
+                            TurbulenceCalibrationPhaseOption.THRESHOLDS,
+                            TurbulenceCalibrationPhaseOption.HISTOGRAM,
+                        ],
+                        calibration_config=TurbulenceCalibrationConfig(
+                            calibration_data_dir=Path("tests/_static/"),
+                            percentile_thresholds=TurbulenceThresholds(
+                                light=0.97, light_to_moderate=98.0, moderate=99.0
+                            ),
+                        ),
+                    )
+                ),
+            ),
+            data_config=DataConfig(
+                spatial_domain=SpatialDomain(
+                    minimum_latitude=-90, maximum_latitude=90, minimum_longitude=-180, maximum_longitude=180
+                )
+            ),
+        )
+
+    return _calibration_only_config
+
+
+class CalibrationOutputFiles(NamedTuple):
+    thresholds: Path | None
+    distribution: Path | None
+
+
+def marry_up_output_files(files: list[Path]) -> CalibrationOutputFiles:
+    match files:
+        case [file_path] if file_path.stem.startswith("thresholds"):
+            return CalibrationOutputFiles(thresholds=files[0], distribution=None)
+        case [file_path] if file_path.stem.startswith("distribution_params"):
+            return CalibrationOutputFiles(thresholds=None, distribution=files[0])
+        case [file_path1, file_path2] if file_path1.stem.startswith("thresholds") and file_path2.stem.startswith(
+            "distribution_params"
+        ):
+            return CalibrationOutputFiles(thresholds=files[0], distribution=files[1])
+        case [file_path1, file_path2] if file_path2.stem.startswith("thresholds") and file_path1.stem.startswith(
+            "distribution_params"
+        ):
+            return CalibrationOutputFiles(thresholds=files[1], distribution=files[0])
+        case _ as unreachable:
[…]
