### 2025-07-22 df5c11a
test: Added integration test which uses the output files from an earlier calibration only run to resume the evaluations
