### 2025-07-22 a735ccc
refactor: Restructure code to prevent ciruclar import when EvaluationDiagnosticsSuite is changed to parent DiagnosticSuite such that this step can be run on both data inputs
