+++ tests/test_configuration.py
@@ -15,6 +15,7 @@ from rojak.orchestrator.configuration import (
     AmdarConfig,
     AmdarDataSource,
     DataConfig,
+    DiagnosticsAmdarHarmonisationStrategyOptions,
     DiagnosticValidationCondition,
     DiagnosticValidationConfig,
     InvalidConfigurationError,
@@ -23,7 +24,6 @@ from rojak.orchestrator.configuration import (
     TurbulenceThresholdMode,
     TurbulenceThresholds,
 )
-from rojak.turbulence.verification import DiagnosticsAmdarHarmonisationStrategyOptions
 from rojak.utilities.types import Limits
 
 if TYPE_CHECKING:
