### 2025-07-22 a5f77aa
refactor: Rework the plumbing so that it is possible to run Amdar verification on calibration data. NOTE: the full plumbing for running on calibration data is not in place yet
