### 2025-07-23 5018e14
refactor: Introduced integration test level conftest which has methods to create the Context to reduce code repetition. It also contains methods to get real data for more comprehensive tests
