### 2025-07-24 f30347a
test: Added test which runs of a larger set of data. The fixture retrieves the netcdf file from the cdsapi then stores the output in a pytest cache.
