### 2025-07-24 01e2cef
fix: Add plots dir so that roc plots are saved in the correct location
