### 2025-07-24 21defc0
fix: Bug fix from running on HPC - no meta was being passed to the map_patitions resulting in it failing
