### 2025-07-24 316541a
ci: Add env vars so that requests to CDS API will work
