### 2025-07-24 8a9c4b8
fix: typo in plots_dir field
