### 2025-07-24 9115b2f
housekeeping: Added logging statements for ROC
