### 2025-07-24 bbff703
fix: Tests were failing as the mock didn't have a Class therby failing the asserts. Thus, passed it the class it should mock with the spec kwarg. Another failure was due to a different method being used to obtain the prototype diagnostic array
