### 2025-07-24 fef2e64
ci: Add type to pull_request event so that workflow does not run on draft PRs
