### 2025-07-24 519421e
fix: Update plot directory to include run name and create the directory for good measure
