+++ src/rojak/core/data.py
@@ -122,7 +122,7 @@ class CATPrognosticData:
         if not set(dataset.coords.keys()).issuperset(self.required_coords):
             missing_coords = self.required_coords - dataset.coords.keys()
             raise ValueError(f"Attempting to instantiate CATPrognosticData with missing coords: {missing_coords}")
-        self._dataset = dataset.persist()
+        self._dataset = dataset.sortby("time").persist()
         blocking_wait_futures(self._dataset)
 
     def temperature(self) -> xr.DataArray:
