### 2025-07-24 754285d
ci: Similarly add event to CD workflow to prevent it running on draft PRs
