+++ .github/workflows/cd.yml
@@ -5,6 +5,8 @@ on:
     branches: [ "master" ]
   pull_request:
     branches: [ "master" ]
+    types:
+      - ready_for_review
   workflow_dispatch:
 
 jobs:
