### 2025-07-24 80b53c0
housekeeping: Added lines to ignore in code coverages. Most important is the type checking block
