+++ pyproject.toml
@@ -99,3 +99,14 @@ packages = ["src/rojak"]
 
 [tool.hatch.version]
 source = "vcs"
+
+[tool.coverage.report]
+exclude_also = [
+    'def __repr__',
+    'raise AssertionError',
+    'raise NotImplementedError',
+    'if 0:',
+    'if __name__ == .__main__.:',
+    'if TYPE_CHECKING:',
+    '@(abc\.)?abstractmethod',
+]
