### 2025-07-25 38cab51
housekeeping: Get coverage to ignore logging statemetns
