+++ pyproject.toml
@@ -109,4 +109,5 @@ exclude_also = [
     'if __name__ == .__main__.:',
     'if TYPE_CHECKING:',
     '@(abc\.)?abstractmethod',
+    '^\s*(logger|logging)\.\w+\(.*\)', # ignore logging statements
 ]
