### 2025-07-25 be76218
fix: don't use minimum step size for ROC calculations
