### 2025-07-25 0b8a170
ci: Add in the default types to pull_requestion statement
