+++ .github/workflows/cd.yml
@@ -6,6 +6,9 @@ on:
   pull_request:
     branches: [ "master" ]
     types:
+      - opened
+      - synchronize
+      - reopened
       - ready_for_review
   workflow_dispatch:
 
