### 2025-07-25 acfa777
ci: Added condition on jobs to prevent running on draft PRs
