### 2025-07-28 a518c2e
fix: set should have two different values where one is EDR
