### 2025-07-28 6071936
feat: Use the fact data is on a regular grid to index into diagnostic_array
