### 2025-07-29 10d2b2b
test: Added assertions to tests
