### 2025-07-30 73bbd68
fix: Error appeared when attempting to use turbulence columns
This fix ensures that the AMDAR turbulence columns are present in the dataframe. They had been unwitting omitted when modifying the existing harmosnisation method into this new implementation.
