### 2025-07-30 c9f1120
fix: Guarantee turbulence diagnostics have been computed
