### 2025-07-30 774a63c
fix: Allow difference to be zero for strictly decreasing
This line checks that the values in the array is decreasing. When that difference is zero, it hits this edge case causing the whole thing to fail. However, if the difference is zero, it is still possible that all the values are decreasing. Therefore, allow it
