### 2025-07-30 2c9c512
fix: Indexing into final value of false_positives threw an error
Invoke persist to create a future for the dask array. This fixed an issue where there was a KeyError for the operation. The stack trace mentioned the future for it had been cancelled. To guarantee that this future is present, persist is inovked to tell the scheduler these values are needed
