### 2025-07-30 e3b33d8
refactor: Added assertions to guarantee columns are in dataframe
