### 2025-07-30 f750f78
feat: Use the nearest value for calculating ROC instead of the bilinear interpolation values
