### 2025-07-30 fca6729
feat: Reduce number of columns loaded from AMDAR data
