### 2025-07-30 d99007b
fix: Invoke persist on variables that need futures pointing to them
