### 2025-07-30 1aa185d
fix: Improve implementation of computing the ROC curve from the neraest diagnotic value
A couple of fixes are in here. 1) level_index was not being saved when adding the nearest diagnostic value. This is needed for the groupby. 2) Use assign for new columns. 3) Add in calls to persist accordingly
