### 2025-07-31 1b9cd9f
feat: Vectorised indexing for diagnostic values
The previous implementation would retrieve the diagnostic value based on the index value stored within the row of the dataframe. When watching the dask distributed dashbaord, I noticed the computation would get stuck on a couple of task never progressing until it magically figures it out. By using the dask.distributed.progress and dask.distributed.print, I managed to narrow down where this bottleneck was occuring.

The bottleneck was in the retrieval of the computed diagnostic value. This location corresponds to the mysterious failure of processi
[…]
