### 2025-07-31 c821a86
refactor: Moved vectorised retrieval of diagnostic values into separate method
