### 2025-07-31 f374bde
refactor: Added comments and move method out into indexing module
Added comments to explain what's being done in _create_diagnostic_value_series so that I understand what it does tmr. Also, move the functionality which changes the order of a list based on another list into a function in indexing. The function also adds additional checks
