+++ src/rojak/core/indexing.py
@@ -167,6 +167,18 @@ def map_values_to_nearest_coordinate_index[T: np.datetime64 | np.number | np.ine
 
 
 def map_order[T](on: list[T], by: list[int]) -> list[T]:
+    """
+    Maps order of ``on`` based on the order specified in ``by``.
+
+    Args:
+        on: List which order is to be mapped onto.
+        by: List which specifies the new order of ``on``
+
+    >>> descending = [5, 4, 3, 2, 1, 0]
+    >>> ascending = list(range(10, 16))
+    >>> map_order(ascending, descending)
+    [15, 14, 13, 12, 11, 10]
+    """
     length: int = len(by)
     by_as_set: set = set(by)
 
@@ -174,10 +186,10 @@ def map_order[T](on: list[T], by: list[int]) -> list[T]:
         raise ValueError("Order mapping must be on lists of the same length")
     if len(by_as_set) != length:
         raise ValueError("Order of items must have unique values")
-    if max(by) != length - 1:
-        raise ValueError(f"Maximum value of the order to impose must be length - 1 not {max(by)}")
-    if min(by) != 0:
-        raise ValueError("Minimum value of the order to impose must be 0")
+    # if max(by) != length - 1:
+    #     raise ValueError(f"Maximum value of the order to impose must be length - 1 not {max(by)}")
+    # if min(by) != 0:
+    #     raise ValueError("Minimum value of the order to impose must be 0")
     if set(range(length)) != by_as_set:
         raise ValueError("Order of items must be increasing by 1 from 0 to length")
 
