### 2025-07-31 efc84b8
test: test for fancy indexing method
