### 2025-07-31 0859cf8
test: Integration test for ROC curve calculation
