### 2025-07-31 6f13c95
fix: Added calls to persist to create futures to values that will be used next
