### 2025-08-01 5e0c367
feat: Make each roc curve unique to which diagnostics were used
