### 2025-08-01 aeb1b77
test: Fixed test to find matching files
File name of ROC curve was updated so that it isn't overwritten by a new run with different turbulence diagnostics. This casued the test to fail as the file name no longer matches that in the test
