### 2025-08-01 36721a3
fix: assign should not have a meta field
This inadvertently created a new column called meta
