### 2025-08-01 7813c7b
fix: Remove meta that should not be in assing
