### 2025-08-01 c8953c7
housekeeping: use np.square ufunc instead of x * x
