### 2025-08-01 f1d287d
housekeeping: replace 1/x with np.reciprocal and another x * x with np.square
