### 2025-08-04 a4ee514
fix: typo in definition of inverse n
