### 2025-08-04 936aaaf
fix: typo when copying pressure from ICAO table
