### 2025-08-04 3289e4f
test: Made tests for the correctness of pressure to altitude conversion stricter
