+++ tests/test_core_calculations.py
@@ -115,9 +115,9 @@ def test_pressure_to_altitude_troposphere_and_vice_versa(wrap_in_data_array: boo
 
     computed_altitude = pressure_to_altitude_troposphere(pressure)
     computed_pressure = altitude_to_pressure_troposphere(altitude_from_table)
-    np.testing.assert_allclose(computed_pressure, pressure, rtol=0.02)
+    np.testing.assert_allclose(computed_pressure, pressure, rtol=1e-5)
     # rtol based on altitude -> pressure as table is from altitude -> pressure
-    np.testing.assert_allclose(computed_altitude, altitude_from_table, rtol=0.02)
+    np.testing.assert_allclose(computed_altitude, altitude_from_table, rtol=1e-5)
     if not is_2d:
         if wrap_in_data_array:
             np.testing.assert_equal(pressure_to_altitude_icao(pressure).values, computed_altitude)  # pyright: ignore[reportAttributeAccessIssue]
@@ -125,8 +125,8 @@ def test_pressure_to_altitude_troposphere_and_vice_versa(wrap_in_data_array: boo
             np.testing.assert_equal(pressure_to_altitude_icao(pressure), computed_altitude)
 
     # Test that we get back approximately the same thing when passed through inverses
-    np.testing.assert_allclose(altitude_to_pressure_troposphere(computed_altitude), pressure, rtol=0.002)
-    np.testing.assert_allclose(pressure_to_altitude_troposphere(computed_pressure), altitude_from_table, rtol=0.002)
+    np.testing.assert_allclose(altitude_to_pressure_troposphere(computed_altitude), pressure)
+    np.testing.assert_allclose(pressure_to_altitude_troposphere(computed_pressure), altitude_from_table)
 
 
 def test_pressure_to_altitude_troposphere_equiv_to_wallace() -> None:
@@ -151,7 +151,8 @@ def test_pressure_to_altitude_stratosphere(is_2d: bool, wrap_in_data_array: bool
         altitude_from_table = xr.DataArray(altitude_from_table)
 
     computed_altitude = pressure_to_altitude_stratosphere(pressure)
-    np.testing.assert_allclose(computed_altitude, altitude_from_table, rtol=0.02)
+    np.testing.assert_equal(np.round(computed_altitude, decimals=-1), altitude_from_table.data)
+    np.testing.assert_allclose(computed_altitude, altitude_from_table, rtol=1e-4)
 
     if not is_2d:
         if wrap_in_data_array:
@@ -192,7 +193,9 @@ def test_pressure_to_altitude_icao(wrap_in_data_array) -> None:
         pressure = xr.DataArray(pressure)
         altitude_from_table = xr.DataArray(altitude_from_table)
 
-    np.testing.assert_allclose(pressure_to_altitude_icao(pressure), altitude_from_table, rtol=0.02)
+    converted_pressure = pressure_to_altitude_icao(pressure)
+    np.testing.assert_equal(np.round(converted_pressure, decimals=-1), altitude_from_table.data)
+    np.testing.assert_allclose(converted_pressure, altitude_from_table, rtol=1e-4)
 
 
 def linear_function(x_vals, y_vals):
