### 2025-08-04 1f460b3
test: fix rtol being too strict and add equal assert
