### 2025-08-04 88a9173
fix: typo in definition of inverse n (#51)
* fix: typo in definition of inverse n

* fix: typo when copying pressure from ICAO table

* test: Made tests for the correctness of pressure to altitude conversion stricter

* test: fix rtol being too strict and add equal assert
