### 2025-08-05 422690d
refactor: Moved method to extract column names into a list into a method
