### 2025-08-05 7e19416
wip: feature: Implemented grid point based AUC aggregation
NOT TESTED OR RUN!! It should work in theory. This is a cleaned up copied over version from the jupyter notebook I was prototyping with. Commit changes cause I will be gutted if I lose this work
