### 2025-08-06 bac0849
feat: Fully implemented and run gridpoint AUC calculation
