### 2025-08-07 b9a2092
improvement: Call cumsum directly on SeriesGroupBy instead of roundabout way
The roundabout way is completely unnecessary. So, I've moved the comment with more details of the complex groupby command to fianlise
