### 2025-08-08 9e2e11e
housekeeping: Commented out unused method and updated data property accordingly
