### 2025-08-08 76e0ce4
feat: Added method to count number of items in group
