### 2025-08-08 981ec97
feat: Implemented alternative method to compute AUC
This method aggregates using the indices instead of the grid box value. This was introduced to see if there was a difference and to hopefully fix the plotting issues I was having
