+++ src/rojak/turbulence/verification.py
@@ -24,7 +24,7 @@ import numpy as np
 import pandas as pd
 
 from rojak.core.calculations import bilinear_interpolation
-from rojak.core.indexing import map_order, map_values_to_nearest_coordinate_index
+from rojak.core.indexing import map_index_to_coordinate_value, map_order, map_values_to_nearest_coordinate_index
 from rojak.orchestrator.configuration import DiagnosticsAmdarHarmonisationStrategyOptions
 from rojak.turbulence.diagnostic import EvaluationDiagnosticSuite
 from rojak.turbulence.metrics import BinaryClassificationResult, area_under_curve, received_operating_characteristic
@@ -656,11 +656,10 @@ class DiagnosticsAmdarVerification:
         return target_data.set_index("multi_index").persist()
 
     def compute_grid_point_auc(
-        self, validation_conditions: "list[DiagnosticValidationCondition]", prototype_diagnostic_array: "xr.DataArray"
+        self, validation_conditions: "list[DiagnosticValidationCondition]", prototype_diagnostic: "xr.DataArray"
     ) -> dict[str, dd.DataFrame]:
-        assert {"pressure_level", "longitude", "latitude", "time"}.issubset(prototype_diagnostic_array.coords)
-        target_data: dd.DataFrame = self._data_harmoniser.nearest_diagnostic_value(self._time_window).persist()
-        space_columns = [self._data_harmoniser.grid_box_column_name, "level_index"]
+        target_data: dd.DataFrame = self.data
+        space_columns = ["lat_index", "lon_index", "level_index"]
         validation_columns = self._get_validation_column_names(validation_conditions)
         target_cols = space_columns + validation_columns + self._data_harmoniser.harmonised_diagnostics
         target_data = target_data[target_cols]
@@ -669,12 +668,15 @@ class DiagnosticsAmdarVerification:
         #    Without a column as the index, the aggregation to compute AUC fails with,
         #       ValueError: If using all scalar values, you must pass an index
         target_data["multi_index"] = (
-            target_data[self._data_harmoniser.grid_box_column_name].astype(str)
+            target_data["lat_index"].astype(str)
+            + "_"
+            + target_data["lon_index"].astype(str)
             + "_"
             + target_data["level_index"].astype(str)
         )
         # Dask doesn't support multi-index so we've got to use a "poor man's" multi-index
         target_data = target_data.set_index("multi_index").persist()
+        # target_data: dd.DataFrame = self._spatial_data_grouping(validation_conditions)
 
         aggregation_spec: dict = {
             condition.observed_turbulence_column_name: condition.grid_point_auc_agg()
@@ -687,7 +689,65 @@ class DiagnosticsAmdarVerification:
             # 2) Sort values so that diagnostic is in descending order as required by the ROC calculation
             #    The set_index() operation performs a sort => need to do sort after and for each diagnostic
             data_for_diagnostic = data_for_diagnostic.sort_values(by=diagnostic_name, ascending=False)
-            data_for_diagnostic = data_for_diagnostic.drop(columns=[diagnostic_name]).persist()
+            data_for_diagnostic = data_for_diagnostic.drop(columns=[diagnostic_name])
+
+            # Specify sort=False to get better performance
+            # See https://docs.dask.org/en/latest/generated/dask.dataframe.DataFrame.groupby.html#dask-dataframe-dataframe-groupby
+            grid_point_auc: dd.DataFrame = data_for_diagnostic.groupby(space_columns, sort=False).aggregate(
+                aggregation_spec, meta={col: pd.Series(dtype=float) for col in validation_columns}
+            )
+
+            grid_point_auc["level_index"] = grid_point_auc.index.map_partitions(
[…]
