### 2025-08-08 3148e49
refactor: Moved checks out into method and fixed line style issue
