### 2025-08-08 1f565fa
feat: Method to plot points from data frame using hvplot
