### 2025-08-08 9db33ae
WIP: Able to plot and call regions AUC from CLI
