### 2025-08-11 014e58f
refactor: Consolidate methods to create plots for polygons or points into a single method
