### 2025-08-11 9743574
refactor: Use AUC that's in the RocVerificationResult instance
