### 2025-08-11 e3e22c6
fix: dictionary must be expanded to kwargs
