+++ src/rojak/plot/turbulence_plotter.py
@@ -497,6 +497,6 @@ def create_interactive_heatmap_plot(
         **(opts_kwargs if opts_kwargs is not None else {}),
     )  # pyright: ignore[reportAssignmentType]
     coast_ls = {"linewidth": 1, "edgecolor": "gray"} if is_matplotlib else {"line_color": "black", "line_width": 0.5}
-    coast: gv.element.geo.Feature = gv.feature.coastline.opts(coast_ls)  # pyright: ignore[reportAssignmentType]
+    coast: gv.element.geo.Feature = gv.feature.coastline.opts(**coast_ls)  # pyright: ignore[reportAssignmentType]
 
     return gv_element * coast
