### 2025-08-12 11fb482
refactor: Specify how to groupby based on a StrEnum and convert logic generalise logic which depends on it
In the original implementation, the logic to use the spatial columns upon which the groupby is performed on is semi-hardcoded. This implementation generalises it so that any list of column names can be passed to it
