+++ src/rojak/turbulence/verification.py
@@ -12,11 +12,13 @@
 #  See the License for the specific language governing permissions and
 #  limitations under the License.
 import dataclasses
+import functools
 import logging
 from abc import ABC, abstractmethod
 from collections import defaultdict
 from collections.abc import Callable, Generator, Mapping
-from typing import TYPE_CHECKING, ClassVar, NamedTuple
+from enum import StrEnum
+from typing import TYPE_CHECKING, ClassVar, NamedTuple, assert_never
 
 import dask.array as da
 import dask.dataframe as dd
@@ -534,6 +536,13 @@ class RocVerificationResult:
         return self._area_under_curve()[amdar_column]
 
 
+class GroupByStrategy(StrEnum):
+    GRID_BOX = "grid_box"
+    GRID_POINT = "grid_point"
+    HORIZONTAL_BOX = "horizontal_box"
+    HORIZONTAL_POINT = "horizontal_point"
+
+
 # Keep this extendable for verification against other forms of data??
 class DiagnosticsAmdarVerification:
     _data_harmoniser: "DiagnosticsAmdarDataHarmoniser"
@@ -553,9 +562,18 @@ class DiagnosticsAmdarVerification:
             return self._harmonised_data
         return self._harmonised_data
 
-    def _grid_spatial_columns(self) -> list[str]:
-        return [self._data_harmoniser.grid_box_column_name, "level_index"]
-        # return ["lat_index", "lon_index", "level_index"]
+    def _grid_spatial_columns(self, group_by_strategy: GroupByStrategy) -> list[str]:
+        match group_by_strategy:
+            case GroupByStrategy.GRID_BOX:
+                return [self._data_harmoniser.grid_box_column_name, "level_index"]
+            case GroupByStrategy.GRID_POINT:
+                return ["lat_index", "lon_index", "level_index"]
+            case GroupByStrategy.HORIZONTAL_BOX:
+                return [self._data_harmoniser.grid_box_column_name]
+            case GroupByStrategy.HORIZONTAL_POINT:
+                return ["lat_index", "lon_index"]
+            case _ as unreachable:
+                assert_never(unreachable)
 
     def _add_nearest_grid_indices(
         self,
@@ -622,24 +640,30 @@ class DiagnosticsAmdarVerification:
             aggregation_spec[strategy_column] = "mean"
         return grouped_by_space_time.aggregate(aggregation_spec)
 
-    def num_obs_per_grid_point(self, validation_conditions: "list[DiagnosticValidationCondition]") -> dd.DataFrame:
-        target_data: dd.DataFrame = self._spatial_data_grouping(validation_conditions)
+    @staticmethod
+    def _get_partition_level_values(partition: pd.Index, level_name: str) -> pd.Index:
+        return partition.get_level_values(level_name)
+
+    def num_obs_per(
+        self, validation_conditions: "list[DiagnosticValidationCondition]", group_by_strategy: GroupByStrategy
+    ) -> dd.DataFrame:
+        target_data: dd.DataFrame = self._spatial_data_grouping(validation_conditions, group_by_strategy)
         turbulence_col: str = validation_conditions[0].observed_turbulence_column_name
-        minimum_columns: list[str] = self._grid_spatial_columns() + [turbulence_col]
-        num_obs = target_data[minimum_columns].groupby(self._grid_spatial_columns()).count()
+        groupby_columns: list[str] = self._grid_spatial_columns(group_by_strategy)
+        minimum_columns: list[str] = groupby_columns + [turbulence_col]
+        num_obs = target_data[minimum_columns].groupby(groupby_columns).count()
 
-        index_right_col: dd.Index = num_obs.index.map_partitions(
-            lambda partition: partition.get_level_values(self._data_harmoniser.grid_box_column_name)
-        )
[…]
