### 2025-08-12 cb93445
refactor: comment out unused method
