+++ src/rojak/turbulence/verification.py
@@ -575,70 +575,70 @@ class DiagnosticsAmdarVerification:
             case _ as unreachable:
                 assert_never(unreachable)
 
-    def _add_nearest_grid_indices(
-        self,
-        validation_columns: list[str],
-        grid_prototype: "xr.DataArray",
-    ) -> "dd.DataFrame":
-        if not set(validation_columns).issubset(self.data.columns):
-            raise ValueError("Validation columns must be present in the data")
+    # def _add_nearest_grid_indices(
+    #     self,
+    #     validation_columns: list[str],
+    #     grid_prototype: "xr.DataArray",
+    # ) -> "dd.DataFrame":
+    #     if not set(validation_columns).issubset(self.data.columns):
+    #         raise ValueError("Validation columns must be present in the data")
+    #
+    #     space_time_columns: list[str] = [
+    #         self._data_harmoniser.common_time_column_name,
+    #         "level",
+    #         "longitude",
+    #         "latitude",
+    #     ]
+    #     target_columns = space_time_columns + validation_columns
+    #     target_data: dd.DataFrame = self.data[target_columns]
+    #     dataframe_meta: dict[str, pd.Series] = get_dataframe_dtypes(target_data)
+    #     dataframe_meta["level_index"] = pd.Series(dtype=int)
+    #     target_data = target_data.map_partitions(
+    #         lambda df: df.assign(
+    #             level_index=df.apply(
+    #                 lambda row, pressure_level=grid_prototype["pressure_level"].values: np.abs(  # noqa: PD011
+    #                     row.level - pressure_level
+    #                 ).argmin(),
+    #                 axis=1,
+    #             )
+    #         ),
+    #         meta=dataframe_meta,
+    #     )
+    #     dataframe_meta["lat_index"] = pd.Series(dtype=int)
+    #     dataframe_meta["lon_index"] = pd.Series(dtype=int)
+    #     return target_data.map_partitions(
+    #         lambda df: df.assign(
+    #             lat_index=map_values_to_nearest_coordinate_index(df.latitude, grid_prototype["latitude"].values),
+    #             lon_index=map_values_to_nearest_coordinate_index(df.longitude, grid_prototype["longitude"].values),
+    #         ),
+    #         meta=dd.from_pandas(pd.DataFrame(dataframe_meta)),
+    #     ).optimize()
 
-        space_time_columns: list[str] = [
-            self._data_harmoniser.common_time_column_name,
-            "level",
-            "longitude",
-            "latitude",
-        ]
-        target_columns = space_time_columns + validation_columns
-        target_data: dd.DataFrame = self.data[target_columns]
-        dataframe_meta: dict[str, pd.Series] = get_dataframe_dtypes(target_data)
-        dataframe_meta["level_index"] = pd.Series(dtype=int)
-        target_data = target_data.map_partitions(
-            lambda df: df.assign(
-                level_index=df.apply(
-                    lambda row, pressure_level=grid_prototype["pressure_level"].values: np.abs(  # noqa: PD011
-                        row.level - pressure_level
-                    ).argmin(),
-                    axis=1,
-                )
-            ),
-            meta=dataframe_meta,
-        )
-        dataframe_meta["lat_index"] = pd.Series(dtype=int)
-        dataframe_meta["lon_index"] = pd.Series(dtype=int)
-        return target_data.map_partitions(
-            lambda df: df.assign(
-                lat_index=map_values_to_nearest_coordinate_index(df.latitude, grid_prototype["latitude"].values),
-                lon_index=map_values_to_nearest_coordinate_index(df.longitude, grid_prototype["longitude"].values),
-            ),
[…]
