### 2025-08-12 f73130c
refactor: Store index column names as properties instead of hard coding them
