### 2025-08-12 b2aab18
refactor: Condensed compute_grid_point_auc and compute_grid_box_auc into a single method
