+++ src/rojak/turbulence/verification.py
@@ -588,13 +588,20 @@ class DiagnosticsAmdarVerification:
     def _grid_spatial_columns(self, group_by_strategy: GroupByStrategy) -> list[str]:
         match group_by_strategy:
             case GroupByStrategy.GRID_BOX:
-                return [self._data_harmoniser.grid_box_column_name, "level_index"]
+                return [
+                    self._data_harmoniser.grid_box_column_name,
+                    self._data_harmoniser.vertical_coordinate_index_column,
+                ]
             case GroupByStrategy.GRID_POINT:
-                return ["lat_index", "lon_index", "level_index"]
+                return [
+                    self._data_harmoniser.latitude_index_column,
+                    self._data_harmoniser.longitude_index_column,
+                    self._data_harmoniser.vertical_coordinate_index_column,
+                ]
             case GroupByStrategy.HORIZONTAL_BOX:
                 return [self._data_harmoniser.grid_box_column_name]
             case GroupByStrategy.HORIZONTAL_POINT:
-                return ["lat_index", "lon_index"]
+                return [self._data_harmoniser.latitude_index_column, self._data_harmoniser.longitude_index_column]
             case _ as unreachable:
                 assert_never(unreachable)
 
@@ -667,6 +674,67 @@ class DiagnosticsAmdarVerification:
     def _get_partition_level_values(partition: pd.Index, level_name: str) -> pd.Index:
         return partition.get_level_values(level_name)
 
+    @staticmethod
+    def _concat_columns_as_str(data_frame: dd.DataFrame, column_names: list[str], separator: str = "_") -> dd.Series:
+        multi_index_columns = [data_frame[col_name].astype(str) for col_name in column_names]
+        return functools.reduce(lambda left, right: left + separator + right, multi_index_columns)
+
+    def _retrieve_grouped_columns(
+        self, data_frame: dd.DataFrame, grouped_columns: list[str], trigger_reset_index: bool
+    ) -> dd.DataFrame:
+        for grouped_column in grouped_columns:
+            data_frame[grouped_column] = data_frame.index.map_partitions(
+                self._get_partition_level_values, grouped_column
+            )
+
+        if trigger_reset_index:
+            data_frame = data_frame.reset_index(drop=True)
+        return data_frame
+
+    def _retrieve_index_column_values(
+        self,
+        data_frame: dd.DataFrame,
+        grouped_columns: list[str],
+        prototype_array: "xr.DataArray",
+        drop_index_cols: bool,
+    ) -> dd.DataFrame:
+        assert {"pressure_level", "longitude", "latitude"}.issubset(prototype_array.coords)
+        index_columns: set[str] = {
+            self._data_harmoniser.longitude_index_column,
+            self._data_harmoniser.latitude_index_column,
+            self._data_harmoniser.vertical_coordinate_index_column,
+        }.intersection(grouped_columns)
+        if not index_columns:  # set is empty
+            return data_frame
+
+        col_name_mapping: dict[str, str] = {
+            self._data_harmoniser.longitude_index_column: "longitude",
+            self._data_harmoniser.latitude_index_column: "latitude",
+            self._data_harmoniser.vertical_coordinate_index_column: "pressure_level",
+        }
+
+        data_frame_dtypes = get_dataframe_dtypes(data_frame)
+        for index_col_name in index_columns:
+            value_col_name = col_name_mapping[index_col_name]
+            data_frame_dtypes[value_col_name] = pd.Series(dtype=float)
+
+        data_frame = data_frame.map_partitions(
+            lambda df: df.assign(
+                **{
+                    col_name_mapping[col_name]: map_index_to_coordinate_value(
[…]
