### 2025-08-12 242bb1f
refactor: renamed class
