### 2025-08-13 df5bdf8
fix: Convert set to list to keep order consistent
