### 2025-08-13 aa7472b
refactor: Moved plotting logic into the turbulence_plotter module
