+++ src/rojak/orchestrator/turbulence.py
@@ -11,9 +11,7 @@
 #  WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
 #  See the License for the specific language governing permissions and
 #  limitations under the License.
-import functools
 import itertools
-import operator
 import sys
 from collections.abc import Iterable, Mapping
 from datetime import datetime
@@ -31,20 +29,16 @@ from rojak.datalib.ukmo.amdar import UkmoAmdarRepository
 from rojak.orchestrator.configuration import (
     AmdarDataSource,
     AmdarDiagnosticCmpSource,
-    DiagnosticsAmdarHarmonisationStrategyOptions,
-    DiagnosticValidationCondition,
     TurbulenceCalibrationPhaseOption,
-    TurbulenceEvaluationConfig,
     TurbulenceEvaluationPhaseOption,
-    TurbulenceEvaluationPhases,
     TurbulenceThresholds,
 )
 from rojak.plot.turbulence_plotter import (
     create_diagnostic_correlation_plot,
-    create_interactive_heatmap_polygon_plot,
+    create_interactive_aggregated_auc_plots,
+    create_interactive_roc_curve_plot,
     create_multi_region_correlation_plot,
     create_multi_turbulence_diagnotics_probability_plot,
-    plot_roc_curve,
     save_hv_plot,
 )
 from rojak.turbulence.analysis import (
@@ -61,24 +55,28 @@ from rojak.turbulence.diagnostic import (
 from rojak.turbulence.verification import (
     DiagnosticsAmdarDataHarmoniser,
     DiagnosticsAmdarVerification,
+    SpatialGroupByStrategy,
 )
 from rojak.utilities.types import DistributionParameters, Limits
 
 if TYPE_CHECKING:
     from pathlib import Path
 
-    import dask.array as da
     import dask.dataframe as dd
 
     from rojak.core.data import AmdarDataRepository, CATData
     from rojak.orchestrator.configuration import Context as ConfigContext
     from rojak.orchestrator.configuration import (
         DataConfig,
+        DiagnosticsAmdarHarmonisationStrategyOptions,
+        DiagnosticValidationCondition,
         SpatialDomain,
         TurbulenceCalibrationConfig,
         TurbulenceCalibrationPhases,
         TurbulenceConfig,
         TurbulenceDiagnostics,
+        TurbulenceEvaluationConfig,
+        TurbulenceEvaluationPhases,
     )
     from rojak.turbulence.verification import (
         RocVerificationResult,
@@ -543,61 +541,43 @@ class DiagnosticsAmdarLauncher:
         if self._validation_conditions:
             logger.info("Starting validation of diagnostics with amdar data")
             verifier = DiagnosticsAmdarVerification(harmoniser, time_window_as_np_datetime)
-            grid_point_auc: dict[str, dd.DataFrame] = verifier.compute_grid_box_auc(self._validation_conditions)
-            num_observations = verifier.num_obs_per_grid_point(self._validation_conditions)
-            # num_pressure_levels: int = diagnostic_suite.get_prototype_computed_diagnostic()["pressure_level"].size
-            target_level = 1
-            num_obs_on_level = num_observations.loc[num_observations["level_index"] == target_level].join(
-                amdar_data.grid, on="index_right"
+            chained_names: str = _chain_diagnostic_names(harmoniser.harmonised_diagnostics)
+            group_by_strategy = SpatialGroupByStrategy.HORIZONTAL_BOX
+            grid_auc = verifier.aggregate_by_auc(
+                self._validation_conditions,
+                diagnostic_suite.get_prototype_computed_diagnostic(),
+                group_by_strategy,
             )
-            for diagnostic_name, regional_auc in grid_point_auc.items():
[…]
