### 2025-08-13 76a52e2
fix: Handle edge case when converting pressure to altitude
When data from ERA5 is used, the pressure_to_altitude_icao method throws a TypeError. This is likely due to the pressure being a coordinate. To fix this, a try catch was instituted and if the exception is thrown, the method is invoked recursively but with a numpy array. This forces any dask back array to a numpy array.

A test was also added to handle this edge case. It is not clear from the committed test but I checked that the TypeError is being thrown when it is raised directly instead of being handled
