### 2025-08-13 00a372f
fix: Change order of multiplication to amke plots look better
