### 2025-08-14 bca7525
feat: Compute central lat lon based on spatial domain
