### 2025-08-14 6e7291b
feat: New colormap to make it easier to differentiate skilled and unskilled AUC scores
