### 2025-08-14 d63ffac
fix: bad logic
