### 2025-08-14 5bad570
refactor: Locally set the extension based on kwarg
