### 2025-08-14 c085a51
housekeeping: make plots bigger
