### 2025-08-15 34b490a
feat: Improve representation of grid boxes with an insufficient amount of observations
In the existing plots, there are two ways in an AUC value of 0 occurs. The first is where there are no positive observations of turbulence. The second is where there is only a single observation at that point. This makes it difficult to interpret the information presented in the plot as a single value maps to multiple meanings. I had been slowly arriving at this conclusion but showing the plots to Ed and Rafa confirmed it.

How it works? The AUC calculation in the group by operation will return np.nan when t
[…]
