### 2025-08-15 f6674e5
feat: 1) two integration method for regional AUC, 2) change where min_group_size is specified
