### 2025-08-15 cc5f8af
refactor: Moved aggregation logic out of configuration
I had always been a bit unsure about the placement of the aggregation logic in the configuration file as it didn't really fit there. But, the condition class lived in there and I felt it should be live with the class so I placed it there.

However, with the refactoring of the aggregation logic into its own method within the metrics, I didn't want to expose that into the configuration module. Moreover, I think it might result in circular imports. This refactoring keeps everything within rojak.turbulence and the aggregation which is only use
[…]
