### 2025-08-18 0431e5f
fix: Ensure dask arrays of dd.Series have the same number of chunks to prevent re-chunking
When running on a larger dataset, e.g. on the HPC, there is a warning in the logs - `PerformanceWarning: Increasing number of chunks by factor of`. This is due to the rechunking of a dask array.

From reading various dask discussions, it appears that this rechunking occurs during invocations of blockwise or map_overlap or map_blocks method. This rechunking is to ensure that the input arrays have the same chunks such that they can be successfully iterated over.

As I know that my arrays will be the same l
[…]
