### 2025-08-18 bf829c4
feat: Function to create different aggregation metrics based on the Enum
