### 2025-08-18 3bff327
feat: Distinguish between points that don't meet the minimum number of observation and those without a positive observation
Curretnly, the aggregation returns two types of invalid values. the first are NaNs which are for grid boxes that don't have enough observations. The second form are those which don't have a positive observation of turbulence. This can cause issues when computing values realted to true positives and false positives as there isn't a real positive.

This drops the first type of invalid values as that informaiton would be captured by the plot showing the distribution of the n
[…]
