### 2025-08-19 4057179
fix: Remove unnecessary drop_vars
The drop_vars is copied over from my prototype impelmentation of the library. It was initially added as it preventing things from running. Removing it still allows it to run on my sample dataset so it has been removed
