### 2025-08-19 c1e49a0
feat: Ensure values of turbIndex column are valid
Values of 63 and 64 are invalid according to the description in the netCDF file. As these values are larger than any valid values, they will be marked as meeting any condition on the intensity of the observed turbulence. This converts any values outside of the valid turbulence to NaNs
