### 2025-08-19 22a6ac5
refactor: Use to_dask_array isntead of .values to make code cleaner and easier to read
