### 2025-08-20 9177dd4
fix: Attempt to eliminate the PerformanceWarning about chunks
The PerformanceWarning is due to the arrays being rechunked at some point. As I can't run the same case on my local machine and it only appears on the HPC jobs it is a bit hard to pin down which line of code is causing it. My best guess is that it is coming from the ROC curve calculations. Thus, to prevent the recalculation of chunks that might lead to different chunking which would later need unifying. This change attempts to compute the chunks at the source of the true positives, false positives and thresholds computation. By comp
[…]
