### 2025-08-20 dbda8d4
fix: Remove unnecessary triggering of diagnostics computation logic
