### 2025-09-09 #8630 @kthyng
@wong-hl
* please move some of the material from your paper to your docs to shorten it (or just remove it)
* see possible missing DOI above
