+++ joss/paper.md
@@ -29,17 +29,10 @@ bibliography: paper.bib
 
 # Summary
 
-<!-- A summary describing the high-level functionality and purpose of the software for a diverse, non-specialist audience.
-
-- [x] Has a clear description of the high-level functionality and purpose of the software for a diverse, non-specialist audience been provided?
-- [x] Explain the research applications of the software -->
-
-<!-- Introduction to the problem/context it exists in -->
 Aviation turbulence is atmospheric turbulence occurring at length scales large enough (approximately 100m to 1km) to affect an aircraft [@sharmanNatureAviationTurbulence2016]. According to the National Transport Safety Board (NTSB), turbulence experienced whilst onboard an aircraft was the leading cause of accidents from 2009 to 2018 [@ntsbPreventingTurbulenceRelatedInjuries2021].
 Clear air turbulence (CAT) is a form of aviation turbulence which cannot be detected by the onboard weather radar. Thus, pilots are unable to preemptively avoid such regions.
 In order to mitigate this safety risk, CAT diagnostics are used to forecast turbulent regions such that pilots are able to tactically avoid them.
 
-<!-- Description of the software -->
 `rojak` is a parallelised python library and command-line tool for using meteorological data to forecast CAT and evaluating the effectiveness of CAT diagnostics against turbulence observations.
 Currently, it supports,
 
@@ -51,41 +44,17 @@ Currently, it supports,
 These features not only allow users to perform operational forecasting of CAT but also to interrogate the intensification in frequency and severity of CAT due to climate change [@williamsIncreasedLightModerate2017; @storerGlobalResponseClearAir2017; @kimGlobalResponseUpperlevel2023], such as by analysing the climatological distribution of the probability of encountering turbulence at different severities (e.g. light turbulence or moderate-or-greater turbulence) for each turbulence diagnostic.
 These applications involve high-volume datasets, ranging from tens to hundreds of gigabytes, necessitating the use of parallelisation to preserve computational tractability and efficiency, while substantially reducing execution time. As such, `rojak` leverages `Dask` to process larger-than-memory data and to run in a distributed manner [@pydataDask].
 
-<!-- Research applications of the software -->
-<!-- In particular, `rojak` opens up the following avenues of research,
-
-1. Exploring whether CAT operational forecasts and our understanding on the effect of climate change on CAT can be improved. This could be achieved by evaluating turbulence diagnostics against observational data to uncover key drivers behind a given incidence of CAT. This could then be used to select the appropriate CAT diagnostic which captures the underlying phenomena. This targeted approach would narrow the uncertainties in the prediction.
-2. Assessing the correlation to other atmospheric features (or phenomena) of interest, such as the likelihood of contrail formation. In the safety-focused aviation industry, it is vital to ensuring the climate-optimised flight trajectories --- which minimise the formation of contrails --- do not result in an increased risk due to turbulence.
-3. Utilising turbulence diagnostics to inform the design of next-generation aircraft and gust load alleviation systems. As the aviation industry moves toward its net-zero goal, future aircraft are expected to be lighter and more aerodynamically efficient. This in turn makes them more susceptible to turbulence. A physics-based model of the intensified CAT enables a more robust aircraft design and the development of effective gust load mitigation strategies. -->
-
 The name of the package, `rojak`, is inspired by its wide range turbulence diagnostics and its applications. While _rojak_ refers to a type of salad, it is also a colloquial term in Malaysia and Singapore for an eclectic mix, reflecting the diverse functionality of the package.
 
 # Statement of need
 
-<!-- A Statement of need section that clearly illustrates the research purpose of the software and places it in the context of related work.
 
[…]
