+++ joss/paper.bib
@@ -36,6 +36,7 @@
   collaborator = {Sharman, Robert D. and Lane, Todd},
   isbn = {978-3-319-23630-8},
   keywords = {Astronautics}
+  doi = {10.1007/978-3-319-23630-8_1}
 }
 
 @techreport{ntsbPreventingTurbulenceRelatedInjuries2021,
