### 2025-09-10 64f8eb1
fix: brokwn doi field
