+++ joss/paper.bib
@@ -35,7 +35,7 @@
   urldate = {2024-07-10},
   collaborator = {Sharman, Robert D. and Lane, Todd},
   isbn = {978-3-319-23630-8},
-  keywords = {Astronautics}
+  keywords = {Astronautics},
   doi = {10.1007/978-3-319-23630-8_1}
 }
 
