### 2025-09-10 379ff4d
fix: missing line which broke figure reference
