### 2025-09-17 93f79ff
feat: Implemented the exporting and loading of computed turbulence diagnostics to zarr files
If zarr version is not limited to being <3, then 1) warning will appear about how the 3.0 specification for strings have not been defined yet and are still under development so there maybe incompatibilities in the future, 2) running the test_turbulence_diagnostics.py::test_to_and_from_zarr will faill due to threads leaked
