+++ pyproject.toml
@@ -32,6 +32,7 @@ dependencies = [
     "scipy",
     "typer",
     "xarray",
+    "zarr>=2,<3", # For outputing diagnostics to file and use zarr 2.0 spec as 3.0 is under development
 ]
 
 [project.urls]
