### 2025-09-24 d685447
fix: Add pyright ignore directive on false positive of da.Array.to_dask_dataframe()
