### 2025-09-25 c911bef
fix: Made the clearing of divisions more robust such that it only clears in the case where divisions are None on the parent data frame
