### 2025-09-25 25cb9b5
fix: Failing test where an edge case meant nchunks != npartitions
