### 2025-09-25 e9ed6a8
fix: pyright directive to ignore attribute access issue was not moved when code was split into multiple lines
