### 2025-09-26 83dcf70
improvement: inline assigning self.data to target_data var
