### 2025-09-26 b6542ec
feat: Require a minimum of two true positive observations to return a valid true and false positive rate
