### 2025-09-26 5976397
fix: Fix broken doctest after updating implemntation to compute lenghts
