### 2025-09-26 1d6e18a
feat: Added functionality to compute the turbulence frequency statistic
This was mainly introduced as a debugging of why the perforamnce in the Southern Hemisphere is so poor. However, I think it should be a mainstay of the code as this is informaiton that will need to be included in a paper
