### 2025-09-26 a2cc6d2
housekeeping: Make code more readable by turning for loop into dictionary comprehension
