### 2025-09-26 cef249e
feat: Put in the plumbing to control the groupby AUC through the configuration file
