### 2025-09-29 8e4db85
fix: Add pyright ignore directive on the dataframe.where conditions
