+++ tests/test_turbulence_mediators.py
@@ -1,225 +1,0 @@
-from typing import TYPE_CHECKING
-
-import numpy as np
-import pytest
-import xarray as xr
-
-from rojak.core.data import AmdarTurbulenceData
-from rojak.orchestrator.configuration import (
-    DiagnosticsAmdarHarmonisationStrategyOptions,
-    TurbulenceDiagnostics,
-    TurbulenceSeverity,
-)
-from rojak.turbulence.diagnostic import DiagnosticSuite, EvaluationDiagnosticSuite
-from rojak.turbulence.verification import (
-    DiagnosticsAmdarDataHarmoniser,
-    DiagnosticsAmdarHarmonisationStrategy,
-    DiagnosticsAmdarHarmonisationStrategyFactory,
-    DiagnosticsSeveritiesStrategy,
-    EdrSeveritiesStrategy,
-    NotWithinTimeFrameError,
-    ValuesStrategy,
-)
-from rojak.utilities.types import Limits
-from tests.conftest import generate_array_data
-
-if TYPE_CHECKING:
-    from pytest_mock import MockerFixture
-
-
-@pytest.mark.parametrize(
-    "time_window",
-    [
-        pytest.param(Limits(np.datetime64("1970-01-01"), np.datetime64("1980-01-01")), id="below_min"),
-        pytest.param(Limits(np.datetime64("2005-02-01T03:00"), np.datetime64("2025-02-01T00:00")), id="above_max"),
-    ],
-)
-def test_diagnostic_amdar_data_harmoniser_execute_harmonisation_fail_time_window(
-    mocker: "MockerFixture", make_dummy_cat_data, time_window: Limits
-) -> None:
-    amdar_data_mock = mocker.Mock(spec=AmdarTurbulenceData)
-    suite_mock = mocker.Mock(spec=DiagnosticSuite)
-    computed_vals_mock = mocker.patch.object(
-        suite_mock, "get_prototype_computed_diagnostic", return_value=make_dummy_cat_data({})
-    )
-
-    harmoniser = DiagnosticsAmdarDataHarmoniser(amdar_data_mock, suite_mock)
-
-    with pytest.raises(NotWithinTimeFrameError) as excinfo:
-        harmoniser.execute_harmonisation([DiagnosticsAmdarHarmonisationStrategyOptions.EDR], time_window)
-
-    assert excinfo.type is NotWithinTimeFrameError
-    computed_vals_mock.assert_called_once()
-
-
-# @pytest.mark.parametrize(
-#     ("grid_shape", "grid_spacing"),
-#     [
-#         pytest.param(box(-130, 25, 28, 60), 0.1, id="grid_spacing_too_small"),
-#         pytest.param(box(-180, 25, 28, 60), 0.25, id="lower_x_limit_too_low"),
-#         pytest.param(box(-130, 0, 28, 60), 0.25, id="lower_y_limit_too_low"),
-#         pytest.param(box(-130, 25, 80, 60), 0.25, id="upper_x_limit_too_high"),
-#         pytest.param(box(-130, 25, 28, 180), 0.25, id="upper_y_limit_too_high"),
-#     ],
-# )
-# def test_diagnostic_amdar_data_harmoniser_execute_harmonisation_fail_on_grid(
-#     grid_shape: "Polygon",
-#     grid_spacing: float,
-#     mocker: "MockerFixture",
-# ) -> None:
-#     x_axis = np.arange(-130, 28, 0.25)
-#     y_axis = np.arange(25, 60, 0.25)
-#     representative_array = xr.DataArray(
-#         data=np.random.default_rng().random((len(x_axis), len(y_axis))),
-#         coords={"longitude": x_axis, "latitude": y_axis},
-#     )
-#     suite_mock = mocker.Mock()
-#     computed_vals_mock = mocker.patch.object(
[…]
