+++ tests/test_core_data.py
@@ -8,11 +8,9 @@ import pandas as pd
 import pytest
 import xarray as xr
 import xarray.testing as xrt
-from shapely.geometry import box
 
-from rojak.core.data import CATData, CATPrognosticData, as_geo_dataframe, expand_grid_bounds
+from rojak.core.data import CATData, CATPrognosticData, as_geo_dataframe
 from rojak.core.derivatives import VelocityDerivative
-from rojak.core.geometric import create_grid_data_frame
 from rojak.datalib.ecmwf.era5 import Era5Data
 from rojak.datalib.madis.amdar import AcarsAmdarRepository
 from rojak.datalib.ukmo.amdar import UkmoAmdarRepository
@@ -513,38 +511,3 @@ def test_convert_to_amdar_turbulence_data(
         "min_lon",
         "max_lon",
     }
-
-
-def test_expand_grid_bounds():
-    grid = create_grid_data_frame(box(10, 10, 12, 12), 1)
-    grid_bounds_df = expand_grid_bounds(grid).compute()
-    bottom_left = {"min_lon": [10], "min_lat": [10], "max_lon": [11], "max_lat": [11]}
-    bottom_right = {"min_lon": [11], "min_lat": [10], "max_lon": [12], "max_lat": [11]}
-    top_left = {"min_lon": [10], "min_lat": [11], "max_lon": [11], "max_lat": [12]}
-    top_right = {"min_lon": [11], "min_lat": [11], "max_lon": [12], "max_lat": [12]}
-    quadrants = [bottom_left, bottom_right, top_left, top_right]
-    for quadrant in quadrants:
-        contains_row: bool = False
-        for row in grid_bounds_df.isin(quadrant).iterrows():
-            # iterrows returns a tuple of [row_number, row]
-            if row[1].all():
-                contains_row = True
-                break
-        assert contains_row
-
-
-def test_join_grid_bounds():
-    grid = create_grid_data_frame(box(10, 10, 12, 12), 1)
-    index_right_values = np.random.default_rng().integers(low=0, high=4, size=100)
-    left_df = dd.from_pandas(pd.DataFrame({"some_value": np.arange(100), "index_right": index_right_values}))
-    repository_instance = AcarsAmdarRepository("")
-    joined_df = repository_instance.join_grid_bounds(left_df, grid).compute()
-    expanded_grid = expand_grid_bounds(grid).compute()
-    for row_num, row in joined_df.iterrows():
-        grid_index = int(row["index_right"])
-        assert grid_index == index_right_values[row_num]
-        pd.testing.assert_series_equal(
-            expanded_grid.iloc[grid_index], row[["min_lon", "min_lat", "max_lon", "max_lat"]], check_names=False
-        )
-
-    assert joined_df.shape == (100, 6)

[…]
