### 2025-09-29 e29cf60
housekeeping: Removed plumbing from CLI for turbuelence harmonisation
Also, updated the test which is to check this functionality of running the harmonisation to expect to fail
