### 2025-09-30 6a54b41
test: Fix test that should have had the min and max of lat and lon removed
