+++ src/rojak/core/data.py
@@ -24,7 +24,6 @@ import dask.array as da
 import dask_geopandas as dgpd
 import distributed
 import numpy as np
-import pandas as pd
 import xarray as xr
 from dask.base import is_dask_collection
 
@@ -306,25 +305,6 @@ def as_geo_dataframe(data_frame: "dd.DataFrame") -> dgpd.GeoDataFrame:
     return gddf.set_crs("epsg:4326")
 
 
-def expand_grid_bounds(grid: "dgpd.GeoDataFrame") -> "dd.DataFrame":
-    def expand_bounds(row: "Polygon") -> pd.Series:
-        min_lon, min_lat, max_lon, max_lat = row.bounds
-        return pd.Series(
-            {
-                "min_lon": np.float64(min_lon),
-                "min_lat": np.float64(min_lat),
-                "max_lon": np.float64(max_lon),
-                "max_lat": np.float64(max_lat),
-            }
-        )
-
-    column_names: list[str] = ["min_lon", "min_lat", "max_lon", "max_lat"]
-    return grid.to_dask_dataframe()["geometry"].apply(
-        expand_bounds,
-        meta=pd.DataFrame({col_name: pd.Series(dtype=np.float64) for col_name in column_names}, dtype=np.float64),
-    )
-
-
 class AmdarDataRepository(ABC):
     """
     Abstract AMDAR data repository interface.
@@ -421,17 +401,6 @@ class AmdarDataRepository(ABC):
     @abstractmethod
     def _time_column_rename_mapping(self) -> dict[str, str]: ...
 
-    def join_grid_bounds(self, dataframe: "dd.DataFrame", grid: "dgpd.GeoDataFrame") -> "dgpd.GeoDataFrame":
-        """
-        Combines data with the grid cell by expanding the Polygon's bounds into 4 columns - min_lon, min_lat,
-        max_lon, max_lat.
-
-        Returns:
-            dd.DataFrame: Combined data frame with the
-
-        """
-        return dataframe.join(expand_grid_bounds(grid), on="index_right", how="left")
-
     def to_amdar_turbulence_data(
         self, target_region: "SpatialDomain | Polygon", grid_size: float, target_pressure_levels: Sequence[float]
     ) -> "AmdarTurbulenceData":
@@ -460,7 +429,6 @@ class AmdarDataRepository(ABC):
 
         grid: dgpd.GeoDataFrame = create_grid_data_frame(target_region, grid_size)
         within_region: dgpd.GeoDataFrame = as_geo_dataframe(raw_data_frame).sjoin(grid).optimize()
-        within_region = self.join_grid_bounds(within_region, grid)
         if self._time_column_rename_mapping():
             within_region = within_region.rename(columns=self._time_column_rename_mapping())
 
@@ -474,7 +442,7 @@ class AmdarTurbulenceData(ABC):
     MINIMUM_ALTITUDE: ClassVar[float] = 8500  # Approx. 28,000 ft
 
     def __init__(self, data_frame: "dd.DataFrame", grid: "dgpd.GeoDataFrame") -> None:
-        required_columns = {"datetime", "index_right", "level", "geometry", "min_lat", "max_lat", "min_lon", "max_lon"}
+        required_columns = {"datetime", "index_right", "level", "geometry"}
         assert required_columns.issubset(data_frame.columns), (
             f"Columns {required_columns - set(data_frame.columns)} from missing the data frame"
         )

[…]
