### 2025-10-03 #8630 @jGaboardi
@wong-hl Before I begin reaching out to potential reviewers, can you comment on the role of your two co-authors for this submission? From the repository there are only code commits made by you, so I'd like to confirm that your co-authors made actual substantive contributions (e.g. manuscript writing, significant ideation, etc.).
