### 2025-10-06 d615967
feat: Method to compute distribution of EDR values
This was introduced as Ko, et al's Evalution and Improvements of the ECMWF Aviation Turbulence Forecasts paper computed the EDR values for the time period they were looking at. As such, I wanted to see if I had comparable values and if I could improve the performance of my diagnsotics using their method
