### 2025-10-06 8a9a94a
WIP: Added classes for jet stream algorithms I want to implement
