### 2025-10-06 6224f9b
feat: NOT TESTED Implemented methodology in Schiemann2009 for identifying the jet
Instead of using the methdology in the python library jsmetrics for finding the local maximas. I have used the second partial derivative test (https://en.wikipedia.org/wiki/Second_partial_derivative_test) for finding the local maximas.

As one of the axis which the derivatives is on (latitude), is affected by the curvature of the earth, a geospatial gradient has been performed to compute the derivative.
