### 2025-10-07 8b9e3a5
feat: NOT TESTED Implemented Koch et al. methodology for identifying a jet stream
The paper uses p_1 and p_2 as 100 and 400 hPa. This has not been strictly enforced in the implementation as the data that is retrieved for CAT analysis doesn't currently use data at those pressure levels. From the descrption in the paper, it sounds like the main reason to why they chose those pressure levels are to capture both the sub-tropical and extratropical jets. If necessary, I will enforced this, but for now, I don't feel that it is needed.

One aspect that was not clear in the paper was which pressure lev
[…]
