### 2025-10-07 9ffbc83
feat: NOT TESTED Implemented method from skimage to find the peaks in a 2D array as using hessian might not be valid
