### 2025-10-08 1c291a9
housekeeping: Removed method to compute local maxima using Hessian
I realised this method is unlikely to work as it compares derivatives to zero. As I need to deal with floats, I will have to decide at what value do we deem it as being close enough to zero that it is zero.
