### 2025-10-08 d8aea0b
fix: After running with real data, velocities were negative, therefore the sign has been flipped
The method has also bee slightly refactored so that the calculation of the alpha velocity field is in a function on its own
