### 2025-10-08 4a119d6
test & docs: Added documentation and tests for get_peak_mask
The tests are a modification of the tests in skimage as the method in skimage returns the coordinates of the local maxima instead of a mask. Moreover, my implementation is a subset of the functionality that's present in the skimage method.
