### 2025-10-08 73637ae
feat: Implemented a factory to instantiated the various jet stream algorithm classes
