### 2025-10-09 f261721
fix: Updated mode for maximum_filter to nearest as grid wrap would not be appropriate for latitudinal-pressure slice
Consider making this a kwarg that is passed in as other modes might be mroe appropriate in other cases
