### 2025-10-09 b02dacf
feat: Method to select values which has been identified as being the jet stream
