### 2025-10-09 b487e21
test: Added peak finding test as I wasn't sure about the behaviour
I wanted to check that when two peaks of equal magnitude are within each other's footprint, it will be flagged as a peak. The second test shows that if one was larger in magnitude to the other, then it is not a local maxima and is not flagged as being a peak
