### 2025-10-09 1f64292
housekeeping: Rename dataarray so that they don't all have the name u_wind
