+++ .github/workflows/release.yml
@@ -33,7 +33,7 @@ jobs:
 #
 #    steps:
 #      - uses: actions/checkout@v5
-#      - uses: astral-sh/setup-uv@v6
+#      - uses: astral-sh/setup-uv@v7
 #      - name: Build wheels
 #        uses: pypa/cibuildwheel@v3.0.1
 #        env:
@@ -53,7 +53,7 @@ jobs:
 
     steps:
       - uses: actions/checkout@v5
-      - uses: astral-sh/setup-uv@v6
+      - uses: astral-sh/setup-uv@v7
       - run: uv build --wheel
       - name: Upload wheel
         uses: actions/upload-artifact@v4
@@ -67,7 +67,7 @@ jobs:
 
     steps:
       - uses: actions/checkout@v5
-      - uses: astral-sh/setup-uv@v6
+      - uses: astral-sh/setup-uv@v7
         with:
           enable-cache: true
           cache-dependency-glob: "uv.lock"
