### 2025-10-13 1864e22
feat: Implemented method to compute matthew's correlation coefficient
