### 2025-10-13 96ad42c
fix: A second PeformanceWarning is thrown at the computation of ROC
This has not been completely eliminated, but has reduced the number of partitoins from increasing by 20+ time to increasing by 12 times
