### 2025-10-13 1fcf21f
refactor: Moved checking whether the confusion matrix has been passed in into a separate function so it can be reused by other methods which need it
