### 2025-10-14 3c8ae3e
test: Fixed failing test due to updated chunking
