### 2025-10-15 827d435
feat: Implemenat mean absolute error metric
