### 2025-10-15 6f7a114
feat: Implemented F1 score metric
