### 2025-10-15 37b8f61
test: Added test for code path where confusion matrix is passed in
