### 2025-10-16 5faeaa7
refactor: Changed implementation of true_skill_score
The original implementation of tss used the output of the ROC curve as that is all I had. However, with the confusion matrix, I now won't need to faff around with the dask array and the code becomes clearer on what it does.
