### 2025-10-16 c195567
test: Fix failing test on Python 3.13 in CI
