+++ tests/integration/test_madis_amdar.py
@@ -23,5 +23,6 @@ def test_climatological_edr(retrieve_single_day_madis_data):
     max_edr: NDArray = acars_data.data_frame["maxEDR"].to_dask_array().compute()
     max_edr = max_edr[max_edr > 0]
     ln_max_edr = np.log(max_edr)
-    np.testing.assert_almost_equal(np.nanmean(ln_max_edr), edr_distribution.mean)
+    # Fails in CI where 7th decimal is different actual is 3, desired is 2
+    np.testing.assert_almost_equal(np.nanmean(ln_max_edr), edr_distribution.mean, decimal=6)
     np.testing.assert_almost_equal(np.nanvar(ln_max_edr), edr_distribution.variance)
