### 2025-10-17 e816765
refactor: Renamed matthews correlation coefficient function to remove typos
