### 2025-10-17 7142beb
docs: Fixed invalid numpy dtype
