### 2025-10-17 0b59d6b
docs: Added theory behind the multidim MCC computation
