### 2025-10-17 7cfe10f
feat: Implemented method to compute contingency table
