### 2025-10-17 f30b6f2
refactor: Use contingency_table method instead computing manually
