### 2025-10-17 954d150
test: Added test to check MCC has same result with equivalent equations
This also checks that the implementation of contingency_table for n_00 is correct
