### 2025-10-17 56aec67
test: Check that n - n_00 == (x | y).sum()
