### 2025-10-17 2af95e7
refactor: Turn repeated code into a pytest fixture
