### 2025-10-17 6f04283
feat: Implemented multidimensional Jaccard Index calculation
Also has tests to check that critical_success_index on 1D is equivalent to Jaccard
