### 2025-10-20 cf115a1
feat: Added class to invoke the matthews correlation coefficient method
