### 2025-10-20 9ace0ba
fix: Added missing __init__ for the child classes
