### 2025-10-20 75276b2
feat: Implemented plumbing to compute relationship between alpha velocity field and turbulence diagnostics
