### 2025-10-21 #8630 @xoolive
Sure count me in
