### 2025-10-21 988200f
WIP: Current iteration to experiment with finding the cases where there is a single jet core
I happily working on this until I realised that I need to re-prioritise for the abstract to get some results related to contrails and tubulence, before working on this
