### 2025-10-21 #9282 @jGaboardi
The PDF [generated without problems during the pre-review](https://github.com/openjournals/joss-reviews/issues/8630#issuecomment-3363796991). @wong-hl , I can't find that you changed anything since then, but please correct me if I am wrong.

cc @kthyng
