+++ .github/workflows/release.yml
@@ -43,7 +43,7 @@ jobs:
 #          CIBW_TEST_COMMAND: >
 #            pytest {project}/tests
 #      - name: Upload wheels
-#        uses: actions/upload-artifact@v4
+#        uses: actions/upload-artifact@v5
 #        with:
 #          name: wheels-${{ matrix.os }}
 #          path: ./dist/*.whl
@@ -56,7 +56,7 @@ jobs:
       - uses: astral-sh/setup-uv@v7
       - run: uv build --wheel
       - name: Upload wheel
-        uses: actions/upload-artifact@v4
+        uses: actions/upload-artifact@v5
         with:
           name: wheel
           path: ./dist/*.whl
@@ -73,7 +73,7 @@ jobs:
           cache-dependency-glob: "uv.lock"
       - run: uv build --sdist
       - name: Upload sdist
-        uses: actions/upload-artifact@v4
+        uses: actions/upload-artifact@v5
         with:
           name: sdist
           path: ./dist/*.tar.gz
