### 2025-10-27 8559f38
refactor: moved contrails to be with jet stream
