### 2025-10-27 78578ea
test: Added tests for exceptions throwns by label_region
