### 2025-10-27 d99cd1e
refactor: Moved checks out into separate function
These checks will need to be reused by a later method. So, make them reusable
