### 2025-10-27 6d9b0c1
docs: Added docstring to methods
