### 2025-10-27 fedb223
refactor: Reduce code repetition by invoking njit method
