+++ src/rojak/atmosphere/regions.py
@@ -116,9 +116,7 @@ def _parent_region_mask(labeled_array: np.ndarray, intersection_mask: np.ndarray
 @guvectorize("void(int32[:, :, :], bool_[:, :, :],bool_[:, :, :])", "(m,n,p),(m,n,p)->(m,n,p)")
 def _parent_region_mask_3d_guv(labeled_array: np.ndarray, intersection_mask: np.ndarray, result: np.ndarray) -> None:
     original_shape = labeled_array.shape
-    flattened_labeled = labeled_array.flatten()
-    target_regions = np.unique(flattened_labeled[intersection_mask.flatten()])
-    out_mask = np.isin(flattened_labeled, target_regions).reshape(original_shape)
+    out_mask = _parent_region_mask(labeled_array, intersection_mask)
     for i in range(original_shape[0]):
         for j in range(original_shape[1]):
             for k in range(original_shape[2]):
@@ -128,9 +126,7 @@ def _parent_region_mask_3d_guv(labeled_array: np.ndarray, intersection_mask: np.
 @guvectorize("void(int32[:, :], bool_[:, :],bool_[:, :])", "(m,n),(m,n)->(m,n)")
 def _parent_region_mask_2d_guv(labeled_array: np.ndarray, intersection_mask: np.ndarray, result: np.ndarray) -> None:
     original_shape = labeled_array.shape
-    flattened_labeled = labeled_array.flatten()
-    target_regions = np.unique(flattened_labeled[intersection_mask.flatten()])
-    out_mask = np.isin(flattened_labeled, target_regions).reshape(original_shape)
+    out_mask = _parent_region_mask(labeled_array, intersection_mask)
     for i in range(original_shape[0]):
         for j in range(original_shape[1]):
             result[i, j] = out_mask[i, j]
