### 2025-10-27 c28baa9
feat: Implement method to combine two features into a single array
