### 2025-10-28 6741b83
feat: Implemented way to compute threhsold based on ROC and TSS (and the plumbing)
For the AIAA paper, I want to use thresholds that are obtained from comparing the forecast and the observations. To do this, I need be able to find the 'best' combination of TP and FP. That's evaluated using the true skill score metric which balances sensitivity and specificity. Using the TSS metric, I find the index which gives the maximum value and index into the thresholds array.

The combination of amdar column and turbulence diagnostic is placed into the dictionary and dumped to file as it's the easiest. Th
[…]
