### 2025-10-29 8990eb2
fix: Equation for true skill score was incorrect
When I plotted this, it didn't look quite right as the max value was just the final value of the array. This uses the definition of true positive rate = sensitivity and false positive rate = 1- specificity. From this chart https://en.wikipedia.org/wiki/Youden%27s_J_statistic#/media/File:ROC_Curve_T1D_Validation,_top5_r2filtered.svg
