### 2025-10-30 affb25c
fix: Corrected equation for TSS
Also added some documentation to explain why the equaiton is the way it is
