### 2025-10-30 73697bf
feat: Include TSS score into plot
As TSS is to be used twice, the computed value is stored in the ROC verification result.
