### 2025-10-30 2cef084
fix: Added which amdar column so they don't overwrite each other
