### 2025-10-31 ImperialCollegeLondon/rojak#73 @danikam
Specify directory in local pip install
In the [installation instructions](https://imperialcollegelondon.github.io/rojak/userguide/installguide.html#installation), the `pip install -e` should also specify the path to the top level of the repo (eg. `pip install -e .`).
