### 2025-10-31 ImperialCollegeLondon/rojak#74 @danikam
Missing python dependencies
[Here in the instructions](https://imperialcollegelondon.github.io/rojak/userguide/getting_started.html#command-line-interface), I need to install the following python dependencies in order for `rojak --help` to run successfully:

```
cdsapi
dask-geopandas
cartopy
pypalettes
```
