### 2025-10-31 1b89ce5
docs: Add missing directory for local pip install
