### 2025-10-31 bc50525
docs: Updated installation instructions to install optional dependencies for CLI
