### 2025-10-31 9df2438
docs: Updated editable install with extra for CLI
