### 2025-10-31 e2cba63
fix: uv sync commands now need an extras flag
I'm being lazy here and just specifying that all extras should be installed
