### 2025-10-31 bd530cd
fix: Remove version specifier on typer
