+++ pyproject.toml
@@ -66,7 +66,7 @@ plot = [
 ]
 cli = [
     "rojak[ecmwf,amdar,plot]",
-    "typer>=0.20.0",
+    "typer",
 ]
 
 [dependency-groups]
