### 2025-10-31 cddb3ff
fix: Using rojak instead of rojak_cat => installed a different dependency NOT this code base
