### 2025-10-31 4bf3607
Fix: Installation from pip is missing optional dependencies  (#76)
* maint: Changed dependency group to optional dependency to make it a package extra

I am unable to read and so I didn't notice that there was a difference between the phrases optional-dependencies and dependency-groups.

optional-dependencies are export as part of the package as an extra, e.g. rojak[cli]. This means that it will be shipped with a release. dependency-groups are not shipped with the package.

* docs: Updated installation instructions to install optional dependencies for CLI

* docs: Updated editable install with
[…]
