### 2025-10-31 #9282 @wong-hl
@danikam I've created a new release on PyPI which ships the optional dependencies required to run `rojak` in the CLI. Running `pip install "rojak-cat[cli]"` will install those dependencies
