### 2025-11-01 ImperialCollegeLondon/rojak#77 @danikam
Clarify target audience in README / user guide
Please clarify the target audience in the README and/or user guide, i.e. who the intended users of the software are or will be.
