### 2025-11-01 ImperialCollegeLondon/rojak#78 @danikam
Add instructions to reproduce functional claims
The Getting Started page in the documentation provides an example yaml file, instructions to retrieve the AMDAR and ERA5 data, and a sample data_config file for the the AMDAR data.

Please additionally include (or point me to) step-by-step instructions in the documentation that a user can follow to run tests that demonstrate the following functional claims in the paper:

* Ability to use meteorological data to compute turbulence diagnostics from ECMWF ERA5 reanalysis.
* How support would be extended to other types of meteorological data
* Compute the 27 turbulence diagnostics
* Convert the turbulence diagnostics values into eddy dissipation rate
* Parallelized execution

If possible (i.e. if the input data file size is reasonable), please also provide step-by-step instructions to reproduce Figures 1 and 2 in the paper.
