### 2025-11-01 #9282 @danikam
https://github.com/ImperialCollegeLondon/rojak/issues/77
https://github.com/ImperialCollegeLondon/rojak/issues/78
