### 2025-11-01 ImperialCollegeLondon/rojak#79 @danikam
Clarify climate change statement
In the [draft paper](https://github.com/ImperialCollegeLondon/rojak/blob/joss-paper/joss/paper.md), please clarify how Figure \autoref{fig:probability_light_turbulence} is showing the impact of climate change (rather than just average probability of encountering light turbulence during the historical period of 2018-2024).
