+++ .pre-commit-config.yaml
@@ -1,7 +1,7 @@
 repos:
   - repo: https://github.com/astral-sh/ruff-pre-commit
     # Ruff version.
-    rev: v0.11.8
+    rev: v0.14.3
     hooks:
       # Run the linter.
       - id: ruff
@@ -10,7 +10,7 @@ repos:
       - id: ruff-format
   - repo: https://github.com/astral-sh/uv-pre-commit
     # uv version.
-    rev: 0.7.8
+    rev: 0.9.7
     hooks:
       - id: uv-lock
   - repo: local
@@ -23,6 +23,6 @@ repos:
         types: [ file ]
         files: \.(rst)$
   - repo: https://github.com/RobertCraigie/pyright-python
-    rev: v1.1.403
+    rev: v1.1.407
     hooks:
       - id: pyright
