### 2025-11-04 92af773
feat: Extended the TransformToEDR class to allow for different c1 and c2 values
The ECMWF methdology would compute the climatological mean and variance for that time period. Using that, map the turbulence diagnostic values to EDR. This extension is to allow for this possibility
