### 2025-11-05 29b5217
feat: Extended TranformToEDR to compute mean and variance in class
The default behaviour is to map the turbulence diagnostic to the mean and variance that was passed in. This allows it to be modified on the fly dpeending on whether the mean and vairance is know ahead of time
