### 2025-11-05 1b754c3
feat: Class to compare a feature with turbulence.
There is a lot of code reptition with the AlphaVel version of this class as they are basically the same thing. So, a refactoring needs to be done. I'll just leave it as it for now as I prepare fro the AIAA forum abstract
