### 2025-11-07 f024e6d
feat: Added more to default colour map
