+++ README.md
@@ -1,4 +1,4 @@
-# rojak : A Library and CLI Tool for Aviation Turbulence Analysis
+# `rojak`: A Library and CLI Tool for Aviation Turbulence Analysis
 
 [![CI](https://github.com/ImperialCollegeLondon/rojak/actions/workflows/ci.yml/badge.svg?branch=master)](https://github.com/ImperialCollegeLondon/rojak/actions/workflows/ci.yml)
 [![CD](https://github.com/ImperialCollegeLondon/rojak/actions/workflows/cd.yml/badge.svg?branch=master)](https://github.com/ImperialCollegeLondon/rojak/actions/workflows/cd.yml)
@@ -8,10 +8,6 @@
 [![Checked with pyright](https://microsoft.github.io/pyright/img/pyright_badge.svg)](https://microsoft.github.io/pyright/)
 [![Formatted with ruff](https://img.shields.io/badge/code%20style-ruff-d7ff64)](https://github.com/astral-sh/ruff)
 [![Contributor Covenant](https://img.shields.io/badge/Contributor%20Covenant-2.1-4baaaa.svg)](CODE_OF_CONDUCT.md)
-
-Rojak is salad with Javanese origins. Colloquially (in Singlish), it means a mixture of things.
-This package is for a traditionally unconventional mixture of aviation meteorology (turbulence diagnostics) and (coming
-soon) aeroelasticity.
 
 `rojak` is a distributed python library and command-line tool for using meterological data to forecast CAT and evaluating the effectiveness of CAT diagnostics against turbulence observations.
 Currently, it supports,
@@ -23,10 +19,15 @@ Currently, it supports,
 
 >[!NOTE]
 > This repository is under active development. As a result,
+>
 > 1. The API is subject to change and may not be stable
 > 2. The documentation is incomplete and will be updated as development progresses.
 >
 > We appreciate your understanding and encourage you to check back for updates.
+
+Rojak is also a salad with Javanese origins. Colloquially (in Singlish), it means a mixture of things.
+This package is for a traditionally unconventional mixture of aviation meteorology (turbulence diagnostics) and (coming
+soon) aeroelasticity.
 
 ## Documentation
 
