### 2025-11-11 a92f185
docs: Added how rojak can be used
Hopefully, by providing example usages of the python package, it'll have clarified the target audience and users
