### 2025-11-11 1fef8d1
docs: shuffled text to make it clearer
