### 2025-11-11 0b47e42
docs: Removed citations from acknowledgments
