### 2025-11-11 ca28e26
docs: Address comment about figure
In #79, @danikam point out that the statement relating to how the figure shows the impact of climate change is unclear. That is 100% spot on, so I've updated it to provide a bit more context
