### 2025-11-11 c6b7f93
maint: Fixed description to reflect it is the current directory not the non-existent input directory
