### 2025-11-11 f0cd186
docs: Copied text from joss paper over to getting started
