### 2025-11-12 32c6de8
wip: Adding in step by step instructions and explanation on how to reproduce results in JOSS paper
