### 2025-11-12 56c9a31
docs: Add config for esbonio vs code plugin for sphinx
