### 2025-11-12 4cd7ef5
wip docs: removed the lines with null values from config
