### 2025-11-12 b920af8
docs: Added note to refer reader to getting started page if the command has failed
