+++ docs/userguide/getting_started.rst
@@ -189,6 +189,8 @@ To use this processed AMDAR data in the turbulence analysis, the :py:class:`roja
           lower: "2024-01-01"
           upper: "2024-12-12T18:00"
 
+.. _met_data:
+
 Meteorological Data: ERA5
 ^^^^^^^^^^^^^^^^^^^^^^^^^^
 
