### 2025-11-13 092b6ee
refactor: Reduce code repetition between the relationship between classes
Also, added in tests to check that they still run after the refacotring. These DO NOT test the correctness only that they run
