### 2025-11-13 101c34c
Revert "MAINT: Update dependencies (#95)" (#96)
This reverts commit 07b511dfa2fab7b25d57c0fe376e9cddd2fa2aeb.
