### 2025-11-14 f676124
test: Added doctest to geodesic methods
Also added some exception handling to ensure inputs are valid
