### 2025-11-14 #9282 @wong-hl
Short update on my progress in addressing the comments

1. Regarding @xoolive's comments
   1. citations: These have been removed in https://github.com/ImperialCollegeLondon/rojak/pull/92
   2. substantial scholarly work: I've put in a bit more detail into the getting started docs to make it clear what this package is (https://github.com/ImperialCollegeLondon/rojak/pull/91) and I'm currently working an examples page to show more functionality beyond the commands to grab data on [this branch](https://github.com/ImperialCollegeLondon/rojak/tree/78-functional-claims)
2. Regarding @danikam's comments / open issues
   1.  I've put in two PRs (https://github.com/ImperialCollegeLondon/rojak/pull/91 and https://github.com/ImperialCollegeLondon/rojak/pull/92) to address two of the issues
   2. For the issue on the functional claims, I'm currently working on the instructions [this branch](https://github.com/ImperialCollegeLondon/rojak/tree/78-functional-claims)
4. The updated word count is 1141. Is that ok? If not, I can make it tighter.
   - I've also added in another affiliation, which bumped up the text outside the main body
