### 2025-11-17 8ab0012
paper: added missing word from danikam's comment
