### 2025-11-18 fe03f63
docs: Implemented a script to compute EDR and plot the snapshot
At the moment, this implements the plotting from scratch. However, the plotting for this has been generalised in the jet-stream branch. Thus, the plan is to run final checks that everthing works on the jet-stream branch. Once that is merged update the script to use that function instead
