### 2025-11-18 bd2be11
refactor: Added kwarg for savefig
Enable bbox_inches and dpi to be configured by whatever calls it instead of forcing it
