### 2025-11-18 a080124
docs: Added descripton to the plot and how to invoke the script
