### 2025-11-18 ec33f86
docs: updated the python command to use uv for esbonio vs code plugin
