+++ pyproject.toml
@@ -129,4 +129,4 @@ exclude_also = [
 
 [tool.esbonio.sphinx]
 buildCommand = ["sphinx-build", "-M", "dirhtml", "docs", "docs/_build"]
-pythonCommand = ["${venv:.venv}"]
+pythonCommand = ["uv", "run", "--group", "docs", "python"]
