### 2025-11-18 8fca98a
refactor: Moved chaining of diagnositcs names together to turbulence plotter
The rationale is that this is really a util functionality related to generating the plot nmaes so it should live in the plotter. It'll also mean that I can easily import this from a place that makes sense into othe rparts of the code
