### 2025-11-18 3499c3d
docs: Updated the python script for edr snapshot
This now uses the API of the methods that were merged in from the jet-stream branch. This makes the plotting a lot cleaner and nicer.
