### 2025-11-18 04d887f
docs: finished instructions for reproducing EDR plot
