### 2025-11-18 a679cb1
docs: renamed file to example uses so I can have example extensions
