### 2025-11-18 cabd687
docs: Added page to give instructions on how to extend rojak to other met sources
