+++ docs/userguide/example-extensions.rst
@@ -1,0 +1,7 @@
+Example Extensions
+====================
+
+Extension to other meteorological data sources
+-------------------------------------------------
+
+lorem ipsum
