### 2025-11-19 f6a64ac
ci: Disable numba for coverage
When numba jit is enabled, coverage isn't able to capture that that code has been tested. This adds whether numba JIT is disabled to the matrix. If it is disabled, then pytest will run with coverage. If it enabled, pytest will run normally.

The uploading of coverage reports to codecov are similarly conditioned to only occur when JIT is disabled
