### 2025-11-20 b44fe77
docs: updated command to specify the pressure levels as defaults are now 6 levels
