### 2025-11-20 8747c08
feat: Added minimal cat contrial default
