### 2025-11-20 45d4bd3
docs: Attempt to fix missing images on deployed docs
