### 2025-11-25 8ee4cb1
docs: Added comment about needing software update to support other met data
