### 2025-11-25 #9282 @wong-hl
@xoolive @danikam I've put together some instructions on the documentation for reproducing the results in the paper. These are in this PR https://github.com/ImperialCollegeLondon/rojak/pull/97. I'm hoping this PR would address all the remaining comments.

I'd recommend checking out to the `78-functional-claims` branch as it has a small bug fix for the EDR colour bar. In addition, it will save you a copying and pasting the python script to create the EDR plots as it can be found in `docs/userguide/edr-snapshot.py`. Do let me know if you encounter any issues as I've not done an end-to-end test of reproducing the results due to how long the data gathering takes.

I'll also be keeping the PR on changes to the JOSS paper (https://github.com/ImperialCollegeLondon/rojak/pull/92) open such that the version that is eventually merged is the accepted version that's close to 1000 words.
