### 2025-11-26 a252dc9
refactor: Moved checks out into separate function to reuse them
