### 2025-11-26 b8eb19a
feat: Implemented method to compute distance from one region to another
This was implemented with the goal of answering the question of how far does one need to move from a contrail forming region to end up in a turbulent region. As this sounded like a solved problem for image analysis, I had a look through scipy to see if there were any available distance methods. Sure enough, these distance transform methods were available.

This uses mathematical morphology to compute the shortest distance from a given point in a feature to another feature. This generalises to any problem that is to find th
[…]
