### 2025-11-26 76b779b
refactor: Reduce code repetition in the preamble and post by passing function and kwargs to another method
