### 2025-11-26 5ecfb1b
refactor: use convenience methods to reduce number of manual args
