### 2025-11-26 795f62e
refactor: Moved repeated functions out into fixtures
