### 2025-11-26 562f6e0
test: Added tests to distance from feature a to feature b  methods
