### 2025-11-26 3292925
docs: Added docstring to methods
