### 2025-11-26 2e06ace
docs: Added doctests from scipy examples
