### 2025-11-27 5de2e45
refactor: Moved generic utility functions out into separate moduel
