+++ src/rojak/plot/turbulence_plotter.py
@@ -13,12 +13,10 @@
 #  limitations under the License.
 
 import functools
-import itertools
 import operator
 import sys
 from collections.abc import Iterable
-from enum import StrEnum
-from typing import TYPE_CHECKING, Final, Literal, assert_never, cast
+from typing import TYPE_CHECKING, Final, Literal, cast
 
 import cartopy.crs as ccrs
 import dask.array as da
@@ -40,6 +38,7 @@ from dask.base import is_dask_collection
 from shapely import Polygon
 
 from rojak.orchestrator.configuration import SpatialDomain, TurbulenceDiagnostics
+from rojak.plot.utilities import _PLATE_CARREE
 from rojak.turbulence.analysis import Hemisphere, LatitudinalRegion
 from rojak.utilities.types import is_dask_array, is_np_array
 
@@ -60,8 +59,6 @@ if TYPE_CHECKING:
     )
     from rojak.turbulence.verification import RocVerificationResult
     from rojak.utilities.types import DiagnosticName
-
-_PLATE_CARREE: "ccrs.Projection" = ccrs.PlateCarree()
 
 GREY_HEX_CODE: str = "#cecece"  # slightly lighter: dbdbdb
 
@@ -230,51 +227,6 @@ def create_multi_turbulence_diagnotics_probability_plot(
     plt.close(fg.fig)
 
 
-class StandardColourMaps(StrEnum):
-    TURBULENCE_PROBABILITY = "turbulence_probability"
-    CORRELATION_COOL_WARM = "correlation_cool_warm"
-    SEQUENTIAL_GREEN = "sequential_green"
-    FEATURE_REGIONS = "feature_regions"
-    DISTANCE_BETWEEN_REGIONS = "distance_between_regions"
-
-
-def get_a_default_cmap(
-    colour_map: StandardColourMaps,
-    resample_to: int | None = None,
-    load_kwargs: dict | None = None,
-) -> "mcolors.Colormap":
-    if load_kwargs is None:
-        load_kwargs = {}
-
-    match colour_map:
-        case StandardColourMaps.TURBULENCE_PROBABILITY:
-            chosen_cmap: mcolors.LinearSegmentedColormap = cast(
-                "mcolors.LinearSegmentedColormap", pypalettes.load_cmap("cancri", cmap_type="continuous", reverse=True)
-            )
-            if resample_to is None:
-                resample_to = 20
-        case StandardColourMaps.DISTANCE_BETWEEN_REGIONS:
-            chosen_cmap: mcolors.LinearSegmentedColormap = cast(
-                "mcolors.LinearSegmentedColormap", pypalettes.load_cmap("cancri", cmap_type="continuous", **load_kwargs)
-            )
-        case StandardColourMaps.CORRELATION_COOL_WARM:
-            chosen_cmap: mcolors.LinearSegmentedColormap = cast(
-                "mcolors.LinearSegmentedColormap", pypalettes.load_cmap("Blue2DarkRed12Steps", **load_kwargs)
-            )
-        case StandardColourMaps.SEQUENTIAL_GREEN:
-            chosen_cmap: mcolors.LinearSegmentedColormap = cast(
-                "mcolors.LinearSegmentedColormap", pypalettes.load_cmap("Ernst", **load_kwargs)
-            )
-        case StandardColourMaps.FEATURE_REGIONS:
-            chosen_cmap: mcolors.LinearSegmentedColormap = cast(
-                "mcolors.LinearSegmentedColormap", pypalettes.load_cmap("Casita1", keep_first_n=4)
-            )
-        case _ as unreachable:
-            assert_never(unreachable)
-
[…]
