### 2025-11-27 d3e780b
feat: Added new module for plotting of regions
Theses functions don't care whether what it is plotting is realted to turbulence
