### 2025-11-27 bba8939
feat: Added in the plotting functions for distances on itnerpolated regions
