+++ src/rojak/plot/regions_plotter.py
@@ -87,6 +87,34 @@ def create_distance_from_a_to_b_instantaneous_lat_lon_plot(  # noqa:PLR0913
     plt.close(fg_dist.fig)
 
 
+def create_distance_from_a_to_b_instantaneous_waypoints_plot(  # noqa: PLR0913
+    distances_from: xr.DataArray,
+    to_regions: xr.DataArray,
+    plot_name: str,
+    time_index: int = 0,
+    waypoint_dim_name: str = "waypoints",
+    vertical_dim_name: str = "pressure_level",
+    plot_kwargs: dict | None = None,
+    savefig_kwargs: dict | None = None,
+) -> None:
+    fig = plt.figure(figsize=(8, 6))
+    ax = fig.add_subplot(1, 1, 1)
+    distances_from.isel(time=time_index).plot(
+        x=waypoint_dim_name,
+        y=vertical_dim_name,
+        yincrease=False,
+        ax=ax,
+        cmap=get_a_default_cmap(StandardColourMaps.DISTANCE_BETWEEN_REGIONS),
+        **(plot_kwargs if plot_kwargs is not None else {}),
+    )
+    to_regions.isel(time=time_index).plot.contour(
+        yincrease=False, ax=ax, levels=1, colors=["red"], linewidths=[0.2], linestyles=["dashed"]
+    )
+    fig.tight_layout()
+    plt.savefig(plot_name, **(savefig_kwargs if savefig_kwargs is not None else {}))
+    plt.close(fig)
+
+
 def create_mean_distances_a_to_b_lat_lon_plot(  # noqa: PLR0913
     distances_from: xr.DataArray,
     plot_name: str,
@@ -97,7 +125,7 @@ def create_mean_distances_a_to_b_lat_lon_plot(  # noqa: PLR0913
     plot_kwargs: dict | None = None,
     savefig_kwargs: dict | None = None,
 ) -> None:
-    fg_mean = distances_from.mean(dim=mean_over_dim).plot(
+    fg_mean = distances_from.mean(dim=mean_over_dim, skipna=True).plot(
         x=lon_coord_name,
         y=lat_coord_name,
         col=column_name,
@@ -111,3 +139,24 @@ def create_mean_distances_a_to_b_lat_lon_plot(  # noqa: PLR0913
     fg_mean.map(lambda: plt.gca().coastlines(lw=0.3))  # pyright: ignore[reportAttributeAccessIssue]
     fg_mean.fig.savefig(plot_name, **(savefig_kwargs if savefig_kwargs is not None else {}))
     plt.close(fg_mean.fig)
+
+
+def create_mean_distances_a_to_b_waypoints_plot(  # noqa: PLR0913
+    distances_from: xr.DataArray,
+    plot_name: str,
+    mean_over_dim: str = "time",
+    waypoint_dim_name: str = "waypoints",
+    vertical_dim_name: str = "pressure_level",
+    plot_kwargs: dict | None = None,
+    savefig_kwargs: dict | None = None,
+) -> None:
+    fig = distances_from.mean(dim=mean_over_dim, skipna=True).plot.contourf(
+        x=waypoint_dim_name,
+        y=vertical_dim_name,
+        yincrease=False,
+        cmap=get_a_default_cmap(StandardColourMaps.DISTANCE_BETWEEN_REGIONS),
+        **(plot_kwargs if plot_kwargs is not None else {}),
+    )
+    fig.tight_layout()
+    plt.savefig(plot_name, **(savefig_kwargs if savefig_kwargs is not None else {}))
+    plt.close(fig)
