### 2025-11-28 3c422dc
maint: Added more linters which has affected the formatting
