+++ src/rojak/orchestrator/configuration.py
@@ -246,7 +246,8 @@ class TurbulenceCalibrationConfig(BaseConfigModel):
         ),
     ] = None
     percentile_thresholds: Annotated[
-        TurbulenceThresholds | None, Field(description="Percentile thresholds", frozen=True, repr=True)
+        TurbulenceThresholds | None,
+        Field(description="Percentile thresholds", frozen=True, repr=True),
     ] = None
     thresholds_file_path: Annotated[
         Path | None,
@@ -257,7 +258,8 @@ class TurbulenceCalibrationConfig(BaseConfigModel):
         ),
     ] = None
     diagnostic_distribution_file_path: Annotated[
-        Path | None, Field(description="Path to directory containing distribution of diagnostic indices")
+        Path | None,
+        Field(description="Path to directory containing distribution of diagnostic indices"),
     ] = None
 
 
@@ -279,14 +281,14 @@ class TurbulenceCalibrationPhases(BaseConfigModel):
                         and self.calibration_config.thresholds_file_path is None
                     ):
                         raise InvalidConfigurationError(
-                            "Either calibration data directory or thresholds file must be provided."
+                            "Either calibration data directory or thresholds file must be provided.",
                         )
                     if (
                         self.calibration_config.calibration_data_dir is not None
                         and self.calibration_config.thresholds_file_path is not None
                     ):
                         raise InvalidConfigurationError(
-                            "Either calibration data directory or thresholds file must be provided, NOT both"
+                            "Either calibration data directory or thresholds file must be provided, NOT both",
                         )
 
                     # Check that the paths are to the correct type of file system object
@@ -307,7 +309,7 @@ class TurbulenceCalibrationPhases(BaseConfigModel):
                         and self.calibration_config.percentile_thresholds is None
                     ):
                         raise InvalidConfigurationError(
-                            "If calibration data directory is provided, threshold configuration must be provided."
+                            "If calibration data directory is provided, threshold configuration must be provided.",
                         )
                 case TurbulenceCalibrationPhaseOption.HISTOGRAM:
                     if (
@@ -359,13 +361,15 @@ class TurbulenceEvaluationConfig(BaseConfigModel):
         Field(description="Target turbulence severity", repr=True),
     ] = [TurbulenceSeverity.LIGHT]
     pressure_levels: Annotated[
-        list[float], Field(description="Pressure levels to evaluate on", repr=True, frozen=True)
+        list[float],
+        Field(description="Pressure levels to evaluate on", repr=True, frozen=True),
     ] = [200.0]
 
 
 class TurbulenceEvaluationPhases(BaseConfigModel):
     phases: Annotated[
-        list[TurbulenceEvaluationPhaseOption], Field(description="Turbulence evaluation phases", repr=True, frozen=True)
+        list[TurbulenceEvaluationPhaseOption],
+        Field(description="Turbulence evaluation phases", repr=True, frozen=True),
     ]
     evaluation_config: TurbulenceEvaluationConfig
 
@@ -380,7 +384,7 @@ class TurbulenceEvaluationPhases(BaseConfigModel):
                     if TurbulenceEvaluationPhaseOption.PROBABILITIES not in self.phases[:index]:
                         raise InvalidConfigurationError(
                             "To compute correlation between probabilities, probabilities phase must be occur before "
-                            "correlations phase"
+                            "correlations phase",
                         )
                 case (
                     TurbulenceEvaluationPhaseOption.CORRELATION_BTW_EDR
@@ -389,7 +393,7 @@ class TurbulenceEvaluationPhases(BaseConfigModel):
                     if TurbulenceEvaluationPhaseOption.EDR not in self.phases[:index]:
[…]
