+++ tests/test_plot_turbulence_plotter.py
@@ -7,11 +7,15 @@ from rojak.plot.turbulence_plotter import _calculate_extent, _cluster_2d_correla
 
 
 @pytest.mark.parametrize(
-    ("min_lat", "max_lat", "min_lon", "max_lon"), [(0, 90, 0, 90), (-90, 90, -180, 180), (30, 45, -10, 20)]
+    ("min_lat", "max_lat", "min_lon", "max_lon"),
+    [(0, 90, 0, 90), (-90, 90, -180, 180), (30, 45, -10, 20)],
 )
 def test_calculate_extent_spatial_domain(min_lat: float, max_lat: float, min_lon: float, max_lon: float) -> None:
     spatial_domain = SpatialDomain(
-        minimum_latitude=min_lat, maximum_latitude=max_lat, minimum_longitude=min_lon, maximum_longitude=max_lon
+        minimum_latitude=min_lat,
+        maximum_latitude=max_lat,
+        minimum_longitude=min_lon,
+        maximum_longitude=max_lon,
     )
     assert _calculate_extent(spatial_domain) == (min_lon, max_lon, min_lat, max_lat)
 
