+++ src/rojak/turbulence/verification.py
@@ -76,7 +76,8 @@ class DiagnosticsAmdarDataHarmoniser:
 
     @staticmethod
     def _check_time_window_within_met_data(
-        time_window: "Limits[np.datetime64]", first_diagnostic: "xr.DataArray"
+        time_window: "Limits[np.datetime64]",
+        first_diagnostic: "xr.DataArray",
     ) -> None:
         time_coordinate: xr.DataArray = first_diagnostic["time"]
         if time_window.lower < time_coordinate.min() or time_window.upper > time_coordinate.max():
@@ -107,7 +108,10 @@ class DiagnosticsAmdarDataHarmoniser:
         return self._diagnostics_suite.diagnostic_names()
 
     def _create_diagnostic_value_series(
-        self, grid_prototype: "xr.DataArray", observational_data: dd.DataFrame, dataframe_meta: dict[str, pd.Series]
+        self,
+        grid_prototype: "xr.DataArray",
+        observational_data: dd.DataFrame,
+        dataframe_meta: dict[str, pd.Series],
     ) -> dict["DiagnosticName", dd.Series]:
         coordinate_axes: list[str] = ["longitude", "latitude", "pressure_level", "time"]
         assert set(coordinate_axes).issubset(grid_prototype.coords)
@@ -146,7 +150,7 @@ class DiagnosticsAmdarDataHarmoniser:
             )
             .persist()
             for diagnostic_name, computed_diagnostic in self._diagnostics_suite.computed_values(
-                "Vectorised indexing to get diagnostic value for observation"
+                "Vectorised indexing to get diagnostic value for observation",
             )
         }
 
@@ -160,8 +164,8 @@ class DiagnosticsAmdarDataHarmoniser:
         observational_data = observational_data.assign(
             level_index=da.abs(
                 observational_data["level"].to_dask_array(lengths=True)[:, np.newaxis]
-                - grid_prototype["pressure_level"].values  # noqa: PD011
-            ).argmin(axis=1)
+                - grid_prototype["pressure_level"].values,  # noqa: PD011
+            ).argmin(axis=1),
         ).persist()
         dataframe_meta[self.latitude_index_column] = pd.Series(dtype=int)
         dataframe_meta[self.longitude_index_column] = pd.Series(dtype=int)
@@ -170,17 +174,19 @@ class DiagnosticsAmdarDataHarmoniser:
             lambda df: df.assign(
                 **{
                     self.latitude_index_column: map_values_to_nearest_coordinate_index(
-                        df.latitude, grid_prototype["latitude"].values
+                        df.latitude,
+                        grid_prototype["latitude"].values,
                     ),
                     self.longitude_index_column: map_values_to_nearest_coordinate_index(
-                        df.longitude, grid_prototype["longitude"].values
+                        df.longitude,
+                        grid_prototype["longitude"].values,
                     ),
                     self.common_time_column_name: map_values_to_nearest_coordinate_index(
                         df.datetime,
                         grid_prototype["time"].values,
                         valid_window=self.TIME_WINDOW_DELTA,
                     ),
-                }
+                },
             ),
             meta=dd.from_pandas(pd.DataFrame(dataframe_meta), npartitions=observational_data.npartitions),
         ).persist()
@@ -230,7 +236,7 @@ def metric_based_aggregation(
             return np.nan
 
         binary_classification: BinaryClassificationRateFromLabels | None = binary_classification_rate_from_cumsum(
-            on_group
+            on_group,
         )
         if binary_classification is None:
             return -1
@@ -244,17 +250,17 @@ def metric_based_aggregation(
             return np.nan
 
         binary_classification: BinaryClassificationRateFromLabels | None = binary_classification_rate_from_cumsum(
-            on_group
+            on_group,
[…]

[…]
