### 2025-11-28 cdd99a9
maint: Change to comply with new RUF linter
