### 2025-11-28 a4bb2eb
maint: Added even more linters, these need to be ignored in the test folder as they will just throw false positives
