+++ ruff.toml
@@ -22,7 +22,7 @@ extend-select = [
     "TID", # Tidy imports
     "TC", # Type checking
     "PT", # Pytest style
-    #    "ARG", # Checks for unused arguments
+    "ARG", # Checks for unused arguments
     "PERF", # Linter for performance anti-patterns
     "PL", # pylint - static code analyser
     "UP", # Upgrade old Python syntax
@@ -33,7 +33,7 @@ extend-select = [
     "PIE",
     "Q",
     "RSE",
-#    "SLF",
+    "SLF",
     "F",
     "FURB",
     "RUF",
@@ -41,7 +41,7 @@ extend-select = [
 ignore = ["ISC001", "COM812"]
 
 [lint.extend-per-file-ignores]
-"tests/*.py" = ["ANN"]
+"tests/*.py" = ["ANN", "SLF", "ARG"]
 
 [lint.pylint]
 max-args = 7
