### 2025-11-28 840116b
feat: Added minimal cat contrial default (#107)
