### 2025-12-03 fc51ef8
refactor: Code path for identifying jet core in both hemisphere is broken so make it NotImplemented
I'm leaning towards making a design decision which only allows for unique jet cores to be identified if the data is for a single hemisphere. This is to simply things especially for the composite stage.
