### 2025-12-03 0c04f94
refactor: Simplified chain statement
