### 2025-12-03 e45f092
refactor: Made unique jet cores a property to reduce recompute
Added result from computing the unique jet core locations to be a value sotred on the instance as it is used multiple times in different functions
