### 2025-12-03 49b95c9
feat: Apply mask to values such that only those within target latitudes are ketp
