### 2025-12-03 71d39d4
feat: Extended region labeller to take in a custom structure
To support the idea I have about compositing only the regions around the jet core, this will allow me to specify a custom shape for the strucutre. For this problem, I want my region to be defined as being contiguous only in the longitudinal direction, thus for a 2D array, I want everything to be false except for the values in the middle column, i.e. struture[:, 1] = True.

This extension gives me the flexibility to specify this custom structure that will not be created by the ndi.generate_binary_structure method
