### 2025-12-04 ac2b8d2
refactor: Shuffled code around and added blank lines
Goal - make it more redable by visually grouping the code together
