### 2025-12-04 66ff4a7
refactor: Moved the mapping of values to index on an irregular grid out into own function as it will be reused
