+++ src/rojak/core/indexing.py
@@ -15,8 +15,10 @@
 from collections.abc import Sequence
 from typing import TYPE_CHECKING
 
+import dask.array as da
 import numpy as np
 import pandas as pd
+from dask.base import is_dask_collection
 
 if TYPE_CHECKING:
     import dask.dataframe as dd
@@ -86,6 +88,14 @@ def get_regular_grid_spacing[T: np.number | np.inexact | np.datetime64 | np.time
             raise NotImplementedError(f"Other dtypes ({array.dtype}) are not yet supported")
 
     return None
+
+
+def map_values_to_nearest_index_irregular_grid(
+    series: "dd.Series | pd.Series", coordinate: "NDArray"
+) -> "da.Array | NDArray":
+    parent_package = da if is_dask_collection(series) else np
+    array: da.Array | NDArray = series.to_dask_array(lengths=True) if is_dask_collection(series) else series.to_numpy()
+    return parent_package.abs(array[:, np.newaxis] - coordinate).argmin(axis=1)
 
 
 def map_values_to_nearest_coordinate_index[T: np.datetime64 | np.number | np.inexact](
