### 2025-12-04 60564f5
refactor: Moved some code out into separate functions as I intedn to reuse them
