### 2025-12-05 30afa17
feat: Removed todense() on sparse array as it works without it.
