### 2025-12-05 6567d3d
docs: Added description to what each kwarg does as it is not obvious
