### 2025-12-05 3ca949d
refactor: Refactored to reduce code repetiton
There is still some repeated code for computing the coordiantes of observations. However, removing that creates issues with how dask parallelises tasks. So, I've accepted it's fine to repeat myself twice.

I've also made it easier to get the grid prototype and the obersvational data. THey are basically wrapper functions. This si as I intedn on introduing a third funvtion which would need to repeat those lines of code

The remaining refactoring is mainly with computing the coordiantes and putting them into the correct order. The axis in which the st
[…]
