### 2025-12-05 07ab8de
feat: Added method to compute total number of gridded observations
