### 2025-12-05 e09de69
refactor: Moved invocation out of list comprehension so it's only invoked once
