### 2025-12-05 5ad3028
test: Added mock for method so tests pass after the refactor
