### 2025-12-05 67fa710
refactor,test: Moved instantiation of DataHarmoniser and its mocks out into a fixture
