### 2025-12-05 ae7b59d
test: Added test for getting the grid coordinates
