### 2025-12-05 ea641f9
test: Added test for counting the number of observations (regridded
