+++ tests/test_turbulence_verification.py
@@ -5,10 +5,12 @@ import dask.dataframe as dd
 import numpy as np
 import pandas as pd
 import pytest
+import xarray as xr
 
 from rojak.core.data import AmdarTurbulenceData
 from rojak.turbulence.diagnostic import DiagnosticSuite
 from rojak.turbulence.verification import DiagnosticsAmdarDataHarmoniser
+from tests.conftest import time_window_dummy_coordinate
 
 if TYPE_CHECKING:
     from pytest_mock import MockerFixture
@@ -175,3 +177,48 @@ def test_create_nearest_diagnostic_value_series_era5_data(
 
     pd.testing.assert_series_equal(values["f3d"].compute(), f3d_values)
     pd.testing.assert_series_equal(values["def"].compute(), def_values)
+
+
+def test_grid_total_observation_count(
+    mocker: "MockerFixture", make_dummy_cat_data, instantiate_diagnostic_amdar_data_harmoniser
+) -> None:
+    harmoniser = instantiate_diagnostic_amdar_data_harmoniser.harmoniser
+    cat_dataset: xr.Dataset = make_dummy_cat_data({}, use_numpy=False)
+
+    index_into_dataset = [
+        {"latitude": 0, "longitude": 2, "pressure_level": 2, "time": 0},
+        {"latitude": 1, "longitude": 5, "pressure_level": 1, "time": 0},
+        {"latitude": 2, "longitude": 8, "pressure_level": 3, "time": 1},
+    ]
+    observational_data = dd.from_pandas(
+        pd.DataFrame(
+            {"lat_index": list(range(3)), "lon_index": [2, 5, 8], "level_index": [2, 1, 3], "time_index": [0, 0, 1]},
+        ),
+        npartitions=2,
+    )
+    coords_of_obs_mock = mocker.patch.object(
+        harmoniser,
+        "_coordinates_of_observations",
+        return_value={
+            "lat_index": observational_data["lat_index"].to_dask_array(lengths=True),
+            "lon_index": observational_data["lon_index"].to_dask_array(lengths=True),
+            "level_index": observational_data["level_index"].to_dask_array(lengths=True),
+            "time_index": observational_data["time_index"].to_dask_array(lengths=True),
+        },
+    )
+
+    desired: xr.DataArray = xr.zeros_like(cat_dataset["geopotential"], dtype=int)
+    for index in index_into_dataset:
+        desired[index] = 1
+
+    suite_mock = instantiate_diagnostic_amdar_data_harmoniser.suite_mock
+    get_prototype_mock = mocker.patch.object(
+        suite_mock, "get_prototype_computed_diagnostic", return_value=cat_dataset["geopotential"]
+    )
+
+    computed = harmoniser.grid_total_observations_count(time_window_dummy_coordinate())
+    get_prototype_mock.assert_called_once()
+    coords_of_obs_mock.assert_called_once()
+
+    computed = computed.data.map_blocks(lambda block: block.todense()).compute()
+    np.testing.assert_array_equal(computed, desired)
