### 2025-12-08 d788a53
docs: Updated incorrect paths
