### 2025-12-08 a114ad0
test: Added test for gridding the frequency in which turbulence is observed
