+++ tests/test_turbulence_verification.py
@@ -5,9 +5,11 @@ import dask.dataframe as dd
 import numpy as np
 import pandas as pd
 import pytest
+import sparse
 import xarray as xr
 
 from rojak.core.data import AmdarTurbulenceData
+from rojak.orchestrator.configuration import DiagnosticValidationCondition
 from rojak.turbulence.diagnostic import DiagnosticSuite
 from rojak.turbulence.verification import DiagnosticsAmdarDataHarmoniser
 from tests.conftest import time_window_dummy_coordinate
@@ -177,6 +179,77 @@ def test_create_nearest_diagnostic_value_series_era5_data(
 
     pd.testing.assert_series_equal(values["f3d"].compute(), f3d_values)
     pd.testing.assert_series_equal(values["def"].compute(), def_values)
+
+
+@pytest.mark.parametrize("min_edr", [0, 0.1, 0.22, 0.5, 1])
+def test_grid_turbulence_observation_frequency_simple(
+    mocker: "MockerFixture", make_dummy_cat_data, instantiate_diagnostic_amdar_data_harmoniser, min_edr: float
+) -> None:
+    harmoniser = instantiate_diagnostic_amdar_data_harmoniser.harmoniser
+    cat_dataset: xr.Dataset = make_dummy_cat_data({}, use_numpy=False)
+
+    index_into_dataset = [
+        {"latitude": 0, "longitude": 2, "pressure_level": 2, "time": 0},
+        {"latitude": 1, "longitude": 5, "pressure_level": 1, "time": 0},
+        {"latitude": 2, "longitude": 8, "pressure_level": 3, "time": 1},
+    ]
+    rand_generator = np.random.default_rng()
+    edr_values = rand_generator.lognormal(-3, 1, size=3)
+    observational_data = dd.from_pandas(
+        pd.DataFrame(
+            {
+                "lat_index": list(range(3)),
+                "lon_index": [2, 5, 8],
+                "level_index": [2, 1, 3],
+                "time_index": [0, 0, 1],
+                "maxEDR": edr_values,
+                "medEDR": edr_values,
+            },
+        ),
+        npartitions=2,
+    )
+    coords_of_obs_mock = mocker.patch.object(
+        harmoniser,
+        "_coordinates_of_observations",
+        return_value={
+            "lat_index": observational_data["lat_index"].to_dask_array(lengths=True),
+            "lon_index": observational_data["lon_index"].to_dask_array(lengths=True),
+            "level_index": observational_data["level_index"].to_dask_array(lengths=True),
+            "time_index": observational_data["time_index"].to_dask_array(lengths=True),
+        },
+    )
+
+    desired: xr.DataArray = xr.zeros_like(cat_dataset["geopotential"], dtype=int)
+    for index_into_array, index_to_dataset in enumerate(index_into_dataset):
+        desired[index_to_dataset] = int(edr_values[index_into_array] > min_edr)
+
+    suite_mock = instantiate_diagnostic_amdar_data_harmoniser.suite_mock
+    get_prototype_mock = mocker.patch.object(
+        suite_mock, "get_prototype_computed_diagnostic", return_value=cat_dataset["geopotential"]
+    )
+    get_observational_data_mock = mocker.patch.object(
+        harmoniser, "_get_observational_data", return_value=observational_data
+    )
+
+    conditions: list[DiagnosticValidationCondition] = [
+        DiagnosticValidationCondition(observed_turbulence_column_name="maxEDR", value_greater_than=min_edr),
+        DiagnosticValidationCondition(observed_turbulence_column_name="medEDR", value_greater_than=min_edr),
+    ]
[…]
