### 2025-12-09 #9282 @jGaboardi
* @danikam
  * very nice
* @xoolive – Once the following PRs are merged, would you be able to give another look through and update [your checklist](https://github.com/openjournals/joss-reviews/issues/9282#issuecomment-3428061210) accordingly?
  * https://github.com/ImperialCollegeLondon/rojak/pull/92
  * https://github.com/ImperialCollegeLondon/rojak/pull/97
