### 2025-12-09 86b3020
Address review comments on JOSS paper  (#92)
* docs: Removed citations from acknowledgments

* docs: Address comment about figure

In #79, @danikam point out that the statement relating to how the figure shows the impact of climate change is unclear. That is 100% spot on, so I've updated it to provide a bit more context

* docs: Added brahmal institute as an affiliation

* paper: added missing word from danikam's comment

* docs: Added comment about needing software update to support other met data
