### 2025-12-09 #9282 @xoolive
@wong-hl few typos:

- line 16 "python" (capitalise)
- line 91 "MeyPy"

otherwise it's all good with me
