### 2025-12-09 #9282 @danikam
Thanks so much @wong-hl for pointing me to the contribution guidelines, they look great from my side!

Minor comments on the updated paper draft:
- line 19: "from European" --> "from the European"
- line 24: "at National" --> "at the National"
- line 51: the ICAO acronym was already introduced on line 32, so may not need to be re-introduced here.
- line 60: "against an earlier time periods" --> "against earlier time periods"
- line 60: "used to characterising" --> "used to characterise"
- line 76: "at 00:00 GMT which" --> "at 00:00 GMT, which"
