+++ tests/test_turbulence_verification.py
@@ -1,3 +1,4 @@
+import warnings
 from collections import namedtuple
 from typing import TYPE_CHECKING
 
@@ -12,7 +13,7 @@ from rojak.core.data import AmdarTurbulenceData
 from rojak.orchestrator.configuration import DiagnosticValidationCondition
 from rojak.turbulence.diagnostic import DiagnosticSuite
 from rojak.turbulence.verification import DiagnosticsAmdarDataHarmoniser
-from tests.conftest import time_window_dummy_coordinate
+from tests.conftest import time_window_dummy_coordinate, time_window_for_cat_data
 
 if TYPE_CHECKING:
     from pytest_mock import MockerFixture
@@ -252,6 +253,88 @@ def test_grid_turbulence_observation_frequency_simple(
     )
 
 
+@pytest.mark.parametrize("min_edr", [0, 0.1, 0.22, 0.5, 1])
+def test_grid_turbulence_observation_frequency_multiple_chunks(
+    mocker: "MockerFixture", load_cat_data, client, instantiate_diagnostic_amdar_data_harmoniser, min_edr: float
+) -> None:
+    suite_mock = instantiate_diagnostic_amdar_data_harmoniser.suite_mock
+    harmoniser: DiagnosticsAmdarDataHarmoniser = instantiate_diagnostic_amdar_data_harmoniser.harmoniser
+    cat_data: CATData = load_cat_data(None, with_chunks=True)
+
+    rand_generator = np.random.default_rng()
+    num_points: int = 30
+    indices: dict[str, np.ndarray] = {
+        "lat_index": rand_generator.integers(0, high=cat_data.u_wind()["latitude"].size, size=num_points),
+        "lon_index": rand_generator.integers(0, high=cat_data.u_wind()["longitude"].size, size=num_points),
+        "level_index": rand_generator.integers(0, high=cat_data.u_wind()["pressure_level"].size, size=num_points),
+        "time_index": rand_generator.integers(0, high=cat_data.u_wind()["time"].size, size=num_points),
+    }
+    index_into_dataset = [
+        {
+            "latitude": indices["lat_index"][index],
+            "longitude": indices["lon_index"][index],
+            "pressure_level": indices["level_index"][index],
+            "time": indices["time_index"][index],
+        }
+        for index in range(num_points)
+    ]
+
+    mean: float = -3
+    std_dev: float = 1
+    indices["maxEDR"] = rand_generator.lognormal(mean=mean, sigma=std_dev, size=num_points)
+    indices["medEDR"] = rand_generator.lognormal(mean=mean, sigma=std_dev, size=num_points)
+
+    observational_data = dd.from_pandas(pd.DataFrame(indices), npartitions=10)
+    coords_of_obs_mock = mocker.patch.object(
+        harmoniser,
+        "_coordinates_of_observations",
+        return_value={
+            "lat_index": observational_data["lat_index"].to_dask_array(lengths=True),
+            "lon_index": observational_data["lon_index"].to_dask_array(lengths=True),
+            "level_index": observational_data["level_index"].to_dask_array(lengths=True),
+            "time_index": observational_data["time_index"].to_dask_array(lengths=True),
+        },
+    )
+
+    get_prototype_mock = mocker.patch.object(
+        suite_mock, "get_prototype_computed_diagnostic", return_value=cat_data.u_wind()
+    )
+    get_observational_data_mock = mocker.patch.object(
+        harmoniser, "_get_observational_data", return_value=observational_data
+    )
+
+    conditions: list[DiagnosticValidationCondition] = [
+        DiagnosticValidationCondition(observed_turbulence_column_name="maxEDR", value_greater_than=min_edr),
+        DiagnosticValidationCondition(observed_turbulence_column_name="medEDR", value_greater_than=min_edr),
+    ]
+
[…]
