### 2025-12-09 881416f
joss: fixed more typos
