+++ joss/paper.md
@@ -39,8 +39,8 @@ In order to mitigate this safety risk, CAT diagnostics are used to forecast turb
 `rojak` is a parallelised Python library and command-line tool for using meteorological data to forecast CAT and evaluating the effectiveness of CAT diagnostics against turbulence observations.
 Currently, it supports,
 
-1. Computing turbulence diagnostics on meteorological data from European Centre for Medium-Range Weather Forecasts's (ECMWF) ERA5 reanalysis on pressure levels [@hersbach2023era5]. Moreover, it is easily extendable through a software update to support other types of meteorological data.
-2. Retrieving and processing turbulence observations from Aircraft Meteorological Data Relay (AMDAR) data archived at National Oceanic and Atmospheric Administration (NOAA)[@NCEPMeteorologicalAssimilation] and AMDAR data collected via the Met Office MetDB system [@ukmoAmdar]
+1. Computing turbulence diagnostics on meteorological data from the European Centre for Medium-Range Weather Forecasts's (ECMWF) ERA5 reanalysis on pressure levels [@hersbach2023era5]. Moreover, it is easily extendable through a software update to support other types of meteorological data.
+2. Retrieving and processing turbulence observations from Aircraft Meteorological Data Relay (AMDAR) data archived at the National Oceanic and Atmospheric Administration (NOAA)[@NCEPMeteorologicalAssimilation] and AMDAR data collected via the Met Office MetDB system [@ukmoAmdar]
 3. Computing 27 different turbulence diagnostics, such as the three-dimensional frontogenesis equation [@bluesteinSynopticdynamicMeteorologyMidlatitudes1993], turbulence index 1 and 2 [@ellrodObjectiveClearAirTurbulence1992], negative vorticity advection [@sharmanIntegratedApproachMid2006], and Brown's Richardson tendency equation [@brownNewIndicesLocate1973].
 4. Converting turbulence diagnostic values into the eddy dissipation rate (EDR) --- the International Civil Aviation Organization's (ICAO) official metric for reporting turbulence [@icaoAnnex3]
 
@@ -52,14 +52,14 @@ The name of the package, `rojak`, is inspired by its wide range turbulence diagn
 # Statement of need
 
 
-Numerous studies have investigated the influence of the climate on CAT [e.g. @kimGlobalResponseUpperlevel2023; @williamsIncreasedLightModerate2017; @storerGlobalResponseClearAir2017] and the application of turbulence diagnostics for operational forecasting [e.g. @gillObjectiveVerificationWorld2014; @sharmanIntegratedApproachMid2006; @sharmanPredictionEnergyDissipation2017; @gillEnsembleBasedTurbulence2014] in products like the Federal Aviation Authority's (FAA) Graphical Turbulence Guidance and the International Civil Aviation Organization's (ICAO) World Area Forecast System.
+Numerous studies have investigated the influence of the climate on CAT [e.g. @kimGlobalResponseUpperlevel2023; @williamsIncreasedLightModerate2017; @storerGlobalResponseClearAir2017] and the application of turbulence diagnostics for operational forecasting [e.g. @gillObjectiveVerificationWorld2014; @sharmanIntegratedApproachMid2006; @sharmanPredictionEnergyDissipation2017; @gillEnsembleBasedTurbulence2014] in products like the Federal Aviation Authority's (FAA) Graphical Turbulence Guidance and the ICAO World Area Forecast System.
 However, to the best of the author's knowledge, none of these studies have made their code publicly available. This work presents the first open-source package for aviation turbulence analysis.
 Given the inherent complexity of CAT diagnostics and the variability in how these could be implemented, `rojak` serves as a first iteration of a standardised implementation of these CAT diagnostics, providing a basis for future enhancements and refinements by the broader research community.
[…]
