### 2025-12-10 3c6336b
joss: Added DOI from zenodo
