### 2025-12-11 #9282 @wong-hl
@jGaboardi I think the DOI on L104 looks that way due to the [URL percent encoding](https://en.wikipedia.org/wiki/Percent-encoding#Character_data) of `<` to `%3C` and `>` to `%3E`
