### 2025-12-11 6319812
WIP fix?: Indexing implementation of gridding the observations
