### 2025-12-11 1515103
WIP refactor: Decoupling amdar data harmoniser from diagnostics
I think my original abstraction and assumption of needing the diagnostic suite is flawed. Thus, I want to have a new class that jsut does the bare minimum the amdar data harmoniser should do. That is to find the coordiantes of the observations, what their index into the grid prototype array is, how many observations there are in total and if there are any positive turublence observations
