+++ src/rojak/turbulence/verification.py
@@ -16,6 +16,7 @@ import functools
 import logging
 from collections import defaultdict
 from collections.abc import Callable, Generator
+from enum import Enum, auto
 from typing import TYPE_CHECKING, ClassVar, Literal, assert_never, cast
 
 import dask.array as da
@@ -87,6 +88,186 @@ def to_coo_array(
     ).compute()
 
     return da.from_array(as_coo, chunks=resulting_chunks)  # pyright: ignore[reportArgumentType]
+
+
+class IndexingFormat(Enum):
+    COORDINATES = auto()
+    FLAT = auto()
+
+
+class AmdarDataHarmoniser:
+    _observational_data: dd.DataFrame
+    _grid_prototype: xr.DataArray
+
+    _time_coord: str
+    _level_coord: str
+    _latitude_coord: str
+    _longitude_coord: str
+    _coordinate_axes: list[str]
+    _time_window_delta: np.timedelta64
+
+    latitude_column: ClassVar[str] = "latitude"
+    longitude_column: ClassVar[str] = "longitude"
+    time_column: ClassVar[str] = "datetime"
+    level_column: ClassVar[str] = "level"
+
+    lat_index_column: ClassVar[str] = "lat_index"
+    lon_index_column: ClassVar[str] = "lon_index"
+    time_index_column: ClassVar[str] = "time_index"
+    level_index_column: ClassVar[str] = "level_index"
+
+    def __init__(  # noqa: PLR0913
+        self,
+        amdar_data: "AmdarTurbulenceData",
+        grid_prototype: xr.DataArray,
+        time_window: "Limits[np.datetime64]",
+        time_coord: str = "time",
+        level_coord: str = "pressure_level",
+        latitude_coord: str = "latitude",
+        longitude_coord: str = "longitude",
+        time_window_delta: np.timedelta64 | None = None,
+    ) -> None:
+        if time_window.lower < grid_prototype[time_coord].min() or time_window.upper > grid_prototype[time_coord].max():
+            raise NotWithinTimeFrameError("Time window is not within time coordinate of met data")
+
+        required_coords: set[str] = {longitude_coord, latitude_coord, time_coord, level_coord}
+        assert required_coords.issubset(grid_prototype.coords)
+        assert required_coords.issubset(grid_prototype.dims)
+        assert grid_prototype.ndim == len(required_coords)
+
+        self._observational_data = amdar_data.clip_to_time_window(time_window).persist()
+        self._grid_prototype = grid_prototype
+
+        self._time_coord = time_coord
+        self._level_coord = level_coord
+        self._latitude_coord = latitude_coord
+        self._longitude_coord = longitude_coord
+        self._coordinate_axes = list(required_coords)
+        if time_window_delta is None:
+            self._time_window_delta = np.timedelta64(3, "h")
+        else:
+            self._time_window_delta = time_window_delta
+
+    def coordinates_of_observations(self) -> dict[str, da.Array]:
+        level_index: da.Array = cast(
+            "da.Array",
+            map_values_to_nearest_index_irregular_grid(
+                self._observational_data[self.level_column], self._grid_prototype[self._level_coord].values
+            ),
+        )
+        latitude_index: da.Array = self._observational_data.map_partitions(
+            lambda df: map_values_to_nearest_coordinate_index(
+                df[self.latitude_column],
+                self._grid_prototype[self._latitude_coord].values,
+            ),
+            # Specifying meta this way resolves this error,
+            #   AttributeError: 'DataFrame' object has no attribute 'name'
+            meta=(self.lat_index_column, int),
+        ).to_dask_array(lengths=True)
+        longitude_index: da.Array = self._observational_data.map_partitions(
+            lambda df: map_values_to_nearest_coordinate_index(
+                df[self.longitude_column],
[…]
