### 2025-12-11 180e9bd
WIP refactor: Return coordiantes in a namedtuple
