### 2025-12-11 7f39cf6
WIP refactor: use the observation_data property instead of private field
