### 2025-12-12 c42b08e
refactor: Removed the usage of the DiagnosticsAmdarDataHarmoniser
The functionality that was achieved by this class has been split into the AmdarDataHarmoniser and the DiagnosticsAmdarVerification this is to generalise the functionality of the classes and to de-couple the class from the DiagnosticSuite.
