+++ src/rojak/turbulence/verification.py
@@ -25,6 +25,7 @@ import numpy as np
 import pandas as pd
 import sparse
 import xarray as xr
+from dask.base import is_dask_collection
 from scipy import integrate
 
 from rojak.core.indexing import (
@@ -173,6 +174,13 @@ class AmdarDataHarmoniser:
     def observational_data(self) -> dd.DataFrame:
         return self._observational_data
 
+    @property
+    def grid_prototype(self) -> xr.DataArray:
+        return self._grid_prototype
+
+    def harmonised_coordinates(self) -> list[str]:
+        return self._coordinate_axes
+
     def coordinates_of_observations(self) -> ObservationCoordinates:
         level_index: da.Array = cast(
             "da.Array",
@@ -293,6 +301,17 @@ class AmdarDataHarmoniser:
             )
 
         return xr.Dataset(data_vars=data_vars, coords=self._grid_prototype.coords)
+
+    def observation_data_with_indices(self) -> dd.DataFrame:
+        coords_of_observation: ObservationCoordinates = self.coordinates_of_observations()
+        return self.observational_data.assign(
+            **{
+                self.level_index_column: coords_of_observation.vertical_index,
+                self.lat_index_column: coords_of_observation.latitude_index,
+                self.lon_index_column: coords_of_observation.latitude_index,
+                self.time_index_column: coords_of_observation.time_index,
+            }
+        )
 
 
 class DiagnosticsAmdarDataHarmoniser:
@@ -759,40 +778,87 @@ class RocVerificationResult:
 
 # Keep this extendable for verification against other forms of data??
 class DiagnosticsAmdarVerification:
-    _data_harmoniser: "DiagnosticsAmdarDataHarmoniser"
+    _data_harmoniser: "AmdarDataHarmoniser"
     _harmonised_data: "dd.DataFrame | None"
-    _time_window: "Limits[np.datetime64]"
+    _turbulence_diagnostics: xr.Dataset
 
-    def __init__(self, data_harmoniser: "DiagnosticsAmdarDataHarmoniser", time_window: "Limits[np.datetime64]") -> None:
+    _grid_box_column: str
+
+    def __init__(
+        self,
+        data_harmoniser: AmdarDataHarmoniser,
+        turbulence_diagnostics: xr.Dataset,
+        grid_box_column: str = "index_right",
+    ) -> None:
+        assert grid_box_column in data_harmoniser.observational_data.columns
+
+        if not set(data_harmoniser.harmonised_coordinates()).issubset(turbulence_diagnostics.coords):
+            raise ValueError(
+                "Coordinates of grid prototype amdar harmonised with does not match turbulence diagnostics"
+            )
+        for harmonised_coord in data_harmoniser.harmonised_coordinates():
+            if not np.issubdtype(turbulence_diagnostics[harmonised_coord].dtype, np.datetime64) and not np.allclose(
+                turbulence_diagnostics[harmonised_coord], data_harmoniser.grid_prototype[harmonised_coord]
+            ):
+                raise ValueError(
+                    "Values of coordinates of grid prototype amdar harmonised are not equal to diagnostics"
+                )
+
+        if not is_dask_collection(turbulence_diagnostics):
+            raise TypeError("Turbulence diagnostics must have dask-backed arrays")
+
         self._data_harmoniser = data_harmoniser
         self._harmonised_data = None
-        self._time_window = time_window
+
+        self._turbulence_diagnostics = turbulence_diagnostics
+        self._grid_box_column = grid_box_column
+
+    def _add_nearest_diagnostic_value(self) -> dd.DataFrame:
+        raveled_index: da.Array = self._data_harmoniser.observations_index_to_grid(IndexingFormat.FLAT).persist()
[…]
