+++ src/rojak/turbulence/diagnostic.py
@@ -1398,6 +1398,9 @@ class DiagnosticSuite:
         _, prototype = next(self.computed_values(""))
         return prototype
 
+    def as_dataset(self) -> xr.Dataset:
+        return xr.Dataset(data_vars=self.computed_values_as_dict())
+
 
 class CalibrationDiagnosticSuite(DiagnosticSuite):
     def __init__(self, factory: DiagnosticFactory, diagnostics: list[TurbulenceDiagnostics]) -> None:
