### 2025-12-12 b994201
fix: Bug in what index was actually being returned by dictionary
