### 2025-12-12 fc08045
test: Added test for getting coordinates from named tuple's dictionary interface
