### 2025-12-12 94e7b51
joss: Fixed broken bibtex
volumne on bluestein should be series isntead
