+++ joss/paper.bib
@@ -392,7 +392,7 @@
   shorttitle = {Synoptic-Dynamic {{Meteorology}} in {{Midlatitudes}}},
   author = {Bluestein, Howard B.},
   year = {1993},
-  volume = {Observations and theory of weather systems},
+  series = {Observations and theory of weather systems},
   publisher = {Taylor \& Francis},
   abstract = {Synoptic meteorology, the study of large-scale weather systems and forecasting using observation, and dynamic meteorology, the study of the laws of physics involved in air movement, are treated in this major new text in two volumes. The author, a meteorologist noted for his research on tornadoes and severe storms, based his work on material he has taught for the past 14 years at the University of Oklahoma. There are no modern texts on the topic. Volume II covers the formation, motion and climatology of extratropical weather systems in the context of the quasigeostrophic theory and "IPV" thinking, the formation and structure of fronts and jets, applications of semigeostrophic theory, and the observed structure and dynamics of precipitation systems in midlatitudes.},
   googlebooks = {AluZNC0Ff2QC},
