### 2025-12-12 86fa9bc
joss: Removed duplicate entry
