+++ joss/paper.bib
@@ -193,21 +193,6 @@
   keywords = {CAT,ensemble,GADS,MOGREPS,ROC,turbulence,verification},
 }
 
-@misc{mayMetPyPythonPackage2016,
-  title = {{{MetPy}}: {{A Python Package}} for {{Meteorological Data}}},
-  shorttitle = {{{MetPy}}},
-  author = {May, Ryan and Arms, Sean and Marsh, Patrick and Bruning, Eric and Leeman, John and Bruick, Zachary and Camron, M. Drew},
-  year = {2016},
-  doi = {10.5065/D6WW7G29},
-  urldate = {2024-12-11},
-  abstract = {MetPy is a collection of tools in Python for reading, visualizing and performing calculations with weather data. The space MetPy aims for is GEMPAK (and maybe NCL)-like functionality, in a way that plugs easily into the existing scientific Python ecosystem (numpy, scipy, matplotlib). So, if you take the average GEMPAK script for a weather map, you need to: - read data - calculate a derived field - show on a map/skew-T One of the benefits hoped to achieve over GEMPAK is to make it easier to use these routines for any meteorological Python application; this means making it easy to pull out the LCL calculation and just use that, or re-use the Skew-T with your own data code. MetPy also prides itself on being well-documented and well-tested, so that on-going maintenance is easily manageable. The intended audience is that of GEMPAK: researchers, educators, and any one wanting to script up weather analysis. It doesn't even have to be scripting; all python meteorology tools are hoped to be able to benefit from MetPy. Conversely, it's hoped to be the meteorological equivalent of the audience of scipy/scikit-learn/skimage.},
-  collaborator = {May, Ryan and University Corporation For Atmospheric Research, Unidata},
-  copyright = {BSD 3-Clause License},
-  howpublished = {UCAR/NCAR - Unidata},
-  langid = {english},
-  keywords = {Data processing,Earth sciences,Free computer software,Geospatial data,Open source software,Python (Computer program language),Software}
-}
-
 @misc{manserSupportDaskArrays,
   title = {Support for {{Dask Arrays}} {$\cdot$} {{Issue}} \#1479 {$\cdot$} {{Unidata}}/{{MetPy}}},
   author = {Manser, Russell},
