### 2025-12-12 4b20e52
fix: Specify what coords to use
Not specifying this meant some test suite cases would fail due to conflicting values on the altitude coord of the data arrays. I think this is due ot how equality is defined and dnoes't accoutn for floats not being exactly equal
