### 2025-12-12 763f11d
fix: Dependencies were occasionally getting lost. Placing persist meant column isnt' lost during/after assign
