### 2025-12-12 0162bff
test: Added tests for getting the index to the grid (coords)
