+++ tests/test_turbulence_verification.py
@@ -14,7 +14,12 @@ import xarray as xr
 from rojak.core.data import AmdarTurbulenceData
 from rojak.orchestrator.configuration import DiagnosticValidationCondition
 from rojak.turbulence.diagnostic import DiagnosticSuite
-from rojak.turbulence.verification import AmdarDataHarmoniser, DiagnosticsAmdarDataHarmoniser, ObservationCoordinates
+from rojak.turbulence.verification import (
+    AmdarDataHarmoniser,
+    DiagnosticsAmdarDataHarmoniser,
+    IndexingFormat,
+    ObservationCoordinates,
+)
 from tests.conftest import time_window_dummy_coordinate, time_window_for_cat_data
 
 if TYPE_CHECKING:
@@ -29,6 +34,8 @@ DataHarmoniserAndMocks = namedtuple(
 
 TestCaseValues = namedtuple("TestCaseValues", ["index_into_dataset", "observational_data"])
 
+AmdarTurbDataMock = namedtuple("AmdarTurbDataMock", ["amdar_data_mock", "observational_data_mock"])
+
 
 @pytest.fixture
 def instantiate_diagnostic_amdar_data_harmoniser(mocker: "MockerFixture") -> DataHarmoniserAndMocks:
@@ -42,6 +49,9 @@ def instantiate_diagnostic_amdar_data_harmoniser(mocker: "MockerFixture") -> Dat
 class TestCaseSize(Enum):
     SMALL = auto()
     LARGE = auto()
+
+
+possible_test_case_sizes: list[TestCaseSize] = [TestCaseSize.SMALL, TestCaseSize.LARGE]
 
 
 @pytest.fixture
@@ -104,7 +114,7 @@ def get_test_case(load_cat_data) -> Callable[[TestCaseSize], TestCaseValues]:
     return _get_test_case
 
 
-@pytest.mark.parametrize("case_size", [TestCaseSize.SMALL, TestCaseSize.LARGE])
+@pytest.mark.parametrize("case_size", possible_test_case_sizes)
 def test_observation_coordinates_as_arrays(get_test_case, case_size: TestCaseSize) -> None:
     case = get_test_case(case_size)
     coords: ObservationCoordinates = ObservationCoordinates(
@@ -125,6 +135,82 @@ def test_observation_coordinates_as_arrays(get_test_case, case_size: TestCaseSiz
         np.testing.assert_array_equal(
             as_arrays[col_name], case.observational_data[col_name].to_dask_array(lengths=True)
         )
+
+
+class TestAmdarDataHarmoniser:
+    @staticmethod
+    def amdar_data_mock(mocker: "MockerFixture", obs_data: "dd.DataFrame") -> AmdarTurbDataMock:
+        amdar_data_mock = mocker.Mock(spec=AmdarTurbulenceData)
+        df_mock = mocker.patch.object(amdar_data_mock, "clip_to_time_window", return_value=obs_data)
+        return AmdarTurbDataMock(amdar_data_mock, df_mock)
+
+    @staticmethod
+    def coords_of_obs_return_value(obs_data: dd.DataFrame) -> ObservationCoordinates:
+        return ObservationCoordinates(
+            obs_data["level_index"].to_dask_array(lengths=True),
+            obs_data["lat_index"],
+            obs_data["lon_index"],
+            obs_data["time_index"],
+        )
+
+    @pytest.mark.parametrize("case_size", possible_test_case_sizes)
+    def test_observations_index_to_grid_fails(
+        self, mocker: "MockerFixture", case_size: TestCaseSize, load_cat_data, get_test_case
+    ) -> None:
+        case: TestCaseValues = get_test_case(case_size)
[…]
