### 2025-12-12 480c140
test: Finished testing code paths in observations_index_to_grid
Also, sneakily added another typeerror check to ensure that axis is not specified when flattening array
