### 2025-12-12 ada3ae8
test: Added test to determine if an observation was present
