### 2025-12-12 621457f
fix,test: Added test and fix for gridding positive turbulence observations
It appears that trying to reduce the number of indices to slcie into created issues. Removing that logic that was supposed to make it more 'efficient' has fixed it and it seems much happier with this implementation
