+++ tests/test_turbulence_verification.py
@@ -32,7 +32,7 @@ DataHarmoniserAndMocks = namedtuple(
     "DataHarmoniserAndMocks", ["harmoniser", "suite_mock", "diagnostic_names_mock", "amdar_data_mock"]
 )
 
-TestCaseValues = namedtuple("TestCaseValues", ["index_into_dataset", "observational_data"])
+TestCaseValues = namedtuple("TestCaseValues", ["index_into_dataset", "observational_data", "indices"])
 
 AmdarTurbDataMock = namedtuple("AmdarTurbDataMock", ["amdar_data_mock", "observational_data_mock"])
 
@@ -66,20 +66,16 @@ def get_test_case(load_cat_data) -> Callable[[TestCaseSize], TestCaseValues]:
                     {"latitude": 2, "longitude": 8, "pressure_level": 0, "time": 1},
                 ]
                 edr_values = rand_generator.lognormal(-3, 1, size=3)
-                observational_data: dd.DataFrame = dd.from_pandas(
-                    pd.DataFrame(
-                        {
-                            "lat_index": list(range(3)),
-                            "lon_index": [2, 5, 8],
-                            "level_index": [2, 1, 0],
-                            "time_index": [0, 0, 1],
-                            "maxEDR": edr_values,
-                            "medEDR": edr_values,
-                        },
-                    ),
-                    npartitions=2,
-                )
-                return TestCaseValues(index_into_dataset, observational_data)
+                indices: dict[str, np.ndarray] = {
+                    "lat_index": np.arange(3),
+                    "lon_index": np.asarray([2, 5, 8]),
+                    "level_index": np.asarray([2, 1, 0]),
+                    "time_index": np.asarray([0, 0, 1]),
+                    "maxEDR": edr_values,
+                    "medEDR": edr_values,
+                }
+                observational_data: dd.DataFrame = dd.from_pandas(pd.DataFrame(indices), npartitions=2)
+                return TestCaseValues(index_into_dataset, observational_data, indices)
             case TestCaseSize.LARGE:
                 cat_data: CATData = load_cat_data(None, with_chunks=True)
                 num_points: int = 30
@@ -107,7 +103,7 @@ def get_test_case(load_cat_data) -> Callable[[TestCaseSize], TestCaseValues]:
                 indices["medEDR"] = rand_generator.lognormal(mean=mean, sigma=std_dev, size=num_points)
 
                 observational_data: dd.DataFrame = dd.from_pandas(pd.DataFrame(indices), npartitions=10)
-                return TestCaseValues(index_into_dataset, observational_data)
+                return TestCaseValues(index_into_dataset, observational_data, indices)
             case _ as unreachable:
                 assert_never(unreachable)
 
@@ -278,6 +274,49 @@ class TestAmdarDataHarmoniser:
         np.testing.assert_array_equal(was_observed, desired)
 
         assert was_observed.sum().compute() == num_points
+
+    @pytest.mark.parametrize("case_size", possible_test_case_sizes)
+    @pytest.mark.parametrize("min_edr", [0, 0.1, 0.22, 0.5, 1])
+    def test_has_positive_turbulence_observation(
+        self, mocker: "MockerFixture", case_size: TestCaseSize, min_edr: float, load_cat_data, get_test_case, client
+    ) -> None:
+        case: TestCaseValues = get_test_case(case_size)
+        amdar_turb_data_mock, _ = self.amdar_data_mock(mocker, case.observational_data)
+
+        cat_data: CATData = load_cat_data(None, with_chunks=True)
+        data_harmoniser: AmdarDataHarmoniser = AmdarDataHarmoniser(
+            amdar_turb_data_mock, cat_data.u_wind(), time_window_for_cat_data()
+        )
+
[…]
