### 2025-12-12 dfc0a2b
test,refactor: moved methods out of class so it can be reused by another
