### 2025-12-12 87337c9
test: Removed repeated tests and ported test for diagnsotic value over to refactored class
