+++ tests/test_turbulence_verification.py
@@ -1,4 +1,3 @@
-import warnings
 from collections import namedtuple
 from collections.abc import Callable
 from enum import Enum, auto
@@ -8,19 +7,17 @@ import dask.dataframe as dd
 import numpy as np
 import pandas as pd
 import pytest
-import sparse
 import xarray as xr
 
 from rojak.core.data import AmdarTurbulenceData
 from rojak.orchestrator.configuration import DiagnosticValidationCondition
-from rojak.turbulence.diagnostic import DiagnosticSuite
 from rojak.turbulence.verification import (
     AmdarDataHarmoniser,
-    DiagnosticsAmdarDataHarmoniser,
+    DiagnosticsAmdarVerification,
     IndexingFormat,
     ObservationCoordinates,
 )
-from tests.conftest import time_window_dummy_coordinate, time_window_for_cat_data
+from tests.conftest import time_window_for_cat_data
 
 if TYPE_CHECKING:
     import dask.array as da
@@ -28,22 +25,10 @@ if TYPE_CHECKING:
 
     from rojak.core.data import CATData
 
-DataHarmoniserAndMocks = namedtuple(
-    "DataHarmoniserAndMocks", ["harmoniser", "suite_mock", "diagnostic_names_mock", "amdar_data_mock"]
-)
 
 TestCaseValues = namedtuple("TestCaseValues", ["index_into_dataset", "observational_data", "indices"])
 
 AmdarTurbDataMock = namedtuple("AmdarTurbDataMock", ["amdar_data_mock", "observational_data_mock"])
-
-
-@pytest.fixture
-def instantiate_diagnostic_amdar_data_harmoniser(mocker: "MockerFixture") -> DataHarmoniserAndMocks:
-    suite_mock = mocker.Mock(spec=DiagnosticSuite)
-    diagnostic_names_mock = mocker.patch.object(suite_mock, "diagnostic_names", return_value=["f3d", "def"])
-    amdar_data_mock = mocker.Mock(spec=AmdarTurbulenceData)
-    harmoniser = DiagnosticsAmdarDataHarmoniser(amdar_data_mock, suite_mock)
-    return DataHarmoniserAndMocks(harmoniser, suite_mock, diagnostic_names_mock, amdar_data_mock)
 
 
 class TestCaseSize(Enum):
@@ -73,6 +58,7 @@ def get_test_case(load_cat_data) -> Callable[[TestCaseSize], TestCaseValues]:
                     "time_index": np.asarray([0, 0, 1]),
                     "maxEDR": edr_values,
                     "medEDR": edr_values,
+                    "index_right": np.arange(3),
                 }
                 observational_data: dd.DataFrame = dd.from_pandas(pd.DataFrame(indices), npartitions=2)
                 return TestCaseValues(index_into_dataset, observational_data, indices)
@@ -101,6 +87,7 @@ def get_test_case(load_cat_data) -> Callable[[TestCaseSize], TestCaseValues]:
                 std_dev: float = 1
                 indices["maxEDR"] = rand_generator.lognormal(mean=mean, sigma=std_dev, size=num_points)
                 indices["medEDR"] = rand_generator.lognormal(mean=mean, sigma=std_dev, size=num_points)
+                indices["index_right"] = np.arange(num_points)
 
                 observational_data: dd.DataFrame = dd.from_pandas(pd.DataFrame(indices), npartitions=10)
                 return TestCaseValues(index_into_dataset, observational_data, indices)
@@ -319,400 +306,39 @@ class TestAmdarDataHarmoniser:
             np.testing.assert_array_equal(computed[col_name], desired_data_vars[col_name])
 
 
[…]
