### 2025-12-12 0efb883
refactor: Removed class that has functionality split between AmdarDataHarmoniser and DiagnosticsAmdarVerification
