+++ src/rojak/turbulence/verification.py
@@ -50,9 +50,8 @@ if TYPE_CHECKING:
 
     from rojak.core.data import AmdarTurbulenceData
     from rojak.orchestrator.configuration import DiagnosticValidationCondition
-    from rojak.turbulence.diagnostic import DiagnosticSuite
     from rojak.turbulence.metrics import BinaryClassificationRateFromLabels
-    from rojak.utilities.types import DiagnosticName, Limits
+    from rojak.utilities.types import Limits
 
 
 logger = logging.getLogger(__name__)
@@ -308,359 +307,6 @@ class AmdarDataHarmoniser:
                 self.lon_index_column: coords_of_observation.latitude_index,
                 self.time_index_column: coords_of_observation.time_index,
             }
-        )
-
-
-class DiagnosticsAmdarDataHarmoniser:
-    """
-    Class brings together the turbulence diagnostics computed from meteorological data and AMDAR observational data
-    to form a unified data set
-    """
-
-    _amdar_data: "AmdarTurbulenceData"
-    _diagnostics_suite: "DiagnosticSuite"
-
-    TIME_WINDOW_DELTA: ClassVar[np.timedelta64] = np.timedelta64(3, "h")
-
-    def __init__(self, amdar_data: "AmdarTurbulenceData", diagnostics_suite: "DiagnosticSuite") -> None:
-        self._amdar_data = amdar_data
-        self._diagnostics_suite = diagnostics_suite
-
-    @staticmethod
-    def _check_time_window_within_met_data(
-        time_window: "Limits[np.datetime64]",
-        first_diagnostic: "xr.DataArray",
-    ) -> None:
-        time_coordinate: xr.DataArray = first_diagnostic["time"]
-        if time_window.lower < time_coordinate.min() or time_window.upper > time_coordinate.max():
-            raise NotWithinTimeFrameError("Time window is not within time coordinate of met data")
-
-    def _get_grid_prototype(self, time_window: "Limits[np.datetime64]") -> "xr.DataArray":
-        grid_prototype: xr.DataArray = self._diagnostics_suite.get_prototype_computed_diagnostic()
-        self._check_time_window_within_met_data(time_window, grid_prototype)
-        return grid_prototype
-
-    def _get_observational_data(self, time_window: "Limits[np.datetime64]") -> dd.DataFrame:
-        return self._amdar_data.clip_to_time_window(time_window).persist()
-
-    @property
-    def common_time_column_name(self) -> str:
-        return "time_index"
-
-    @property
-    def grid_box_column_name(self) -> str:
-        return "index_right"
-
-    @property
-    def latitude_index_column(self) -> str:
-        return "lat_index"
-
-    @property
-    def longitude_index_column(self) -> str:
-        return "lon_index"
-
-    @property
-    def vertical_coordinate_index_column(self) -> str:
-        return "level_index"
-
-    @property
-    def coordinate_axes(self) -> list[str]:
-        return ["longitude", "latitude", "pressure_level", "time"]
-
-    @property
-    def index_column_names(self) -> list[str]:
-        return [
-            self.longitude_index_column,
-            self.latitude_index_column,
-            self.vertical_coordinate_index_column,
-            self.common_time_column_name,
-        ]
-
-    def coordinate_axes_to_index_col_name(self) -> dict[str, str]:
-        return dict(zip(self.coordinate_axes, self.index_column_names, strict=True))
-
-    @property
-    def harmonised_diagnostics(self) -> list[str]:
-        return self._diagnostics_suite.diagnostic_names()
-
-    def _get_gridded_coordinates(
-        self, observational_data: dd.DataFrame, grid_prototype: xr.DataArray, stack_on_axis: int
-    ) -> da.Array:
[…]
