### 2025-12-12 1991d4b
test: Added an additional assert to check total number of positive observations is correct
