### 2025-12-12 a462fae
test,fix: Added test for num_obs_per which highlighted a tokenization issue that has beenf ixed
