### 2025-12-12 f357c11
test: Added test for aggregate_by_auc method runs
