### 2025-12-12 9f2ac51
refactor: Removed unused code
