### 2025-12-15 91341fd
test: Added test for nearest_value_roc method
