### 2025-12-15 0d770f9
test: Convert desired answer to a Python float so that actual and desired are both Python floats
This test was failing after the xarray version was updated and when upgrading the pytest mock version. This was due to the conversion of a numpy float to a python float resulting in some errors. Thus, making a numpy float not almost equal to a python float
