### 2025-12-15 e27a486
test: Attempt to fix failing test in CI
For some reason, the tests fail in the CI with a h5 related error, so I'm going to try fixing it by changing the default engine. I can't replicate this failure mode locally.
