+++ tests/integration/test_madis_amdar.py
@@ -33,7 +33,7 @@ def test_climatological_edr(retrieve_single_day_madis_data):
     max_edr = max_edr[max_edr > 0]
     ln_max_edr = np.log(max_edr)
     # Fails in CI where 7th decimal is different actual is 3, desired is 2
-    np.testing.assert_almost_equal(edr_distribution.mean, float(np.nanmean(ln_max_edr)))
+    np.testing.assert_almost_equal(edr_distribution.mean, float(np.nanmean(ln_max_edr)), decimal=6)
     np.testing.assert_almost_equal(edr_distribution.variance, float(np.nanvar(ln_max_edr)))
 
 
