### 2025-12-16 d4abd8a
refactor: Moved climatological EDR values to rojak.core.constants modeule
