### 2025-12-16 a321144
feat: Added climatological EDR dist for 2020 to 2024
