### 2025-12-16 018833a
refactor: Renamed climatological parameter s to make it clear it is from sharman 2017
