+++ src/rojak/turbulence/analysis.py
@@ -27,7 +27,7 @@ from pydantic.dataclasses import dataclass as pydantic_dataclass
 from rich.progress import track
 
 from rojak.core.analysis import PostProcessor
-from rojak.core.constants import OVERALL_CLIMATOLOGICAL_PARAMETER
+from rojak.core.constants import SHARMAN_17_CLIMATOLOGICAL_PARAMETER
 from rojak.orchestrator.configuration import (
     RelationshipBetweenTypes,
     TurbulenceSeverity,
@@ -366,8 +366,8 @@ class TransformToEDR(PostProcessor):
             self._c1 = c1
             self._c2 = c2
         elif c1 is None and c2 is None:
-            self._c1 = OVERALL_CLIMATOLOGICAL_PARAMETER.c1
-            self._c2 = OVERALL_CLIMATOLOGICAL_PARAMETER.c2
+            self._c1 = SHARMAN_17_CLIMATOLOGICAL_PARAMETER.c1
+            self._c2 = SHARMAN_17_CLIMATOLOGICAL_PARAMETER.c2
         else:
             raise TypeError("Both c1 and c2 must be both be provided or omitted")
 
