### 2025-12-16 753a5d0
FEAT: Add climatological EDR distribution of observations from 2020-2024 (#127)
* refactor: Moved climatological EDR values to rojak.core.constants modeule

* feat: Added climatological EDR dist for 2020 to 2024

* refactor: Renamed climatological parameter s to make it clear it is from sharman 2017
