### 2026-01-05 d7aece6
refactor: Update how kwargs are specified and what is returend so it works in a jupyter notebook with customisations
