### 2026-01-05 dd5d1a1
maint: Added precommit-hooks which check file format, eol and trailng spaces => formatting changes
