### 2026-01-05 6f50df9
MAINT: Add Python 3.14 as a supported version and to test suite (#133)
* docs: Added supported python version badge

* ci: Added python 3.14 to the test suite
