### 2026-01-06 b08778d
MAINT: Add pre commit hooks (#134)
* maint: Added precommit-hooks which check file format, eol and trailng spaces => formatting changes

* fix: Attempt to force compute to fix CI failures
