+++ docs/_templates/autosummary/class.rst
@@ -30,5 +30,3 @@
    {%- endfor %}
    {% endif %}
    {% endblock %}
-
-
