### 2026-01-14 385124c
refactor: Added Final type hint to constants
