### 2026-01-14 1f2ece6
refactor: Rename class to prevent class with PytestCollectionWarning about TestCaseValues
See error message from CI

tests/test_turbulence_verification.py:39
  /home/runner/work/rojak/rojak/tests/test_turbulence_verification.py:39: PytestCollectionWarning: cannot collect test class 'TestCaseSize' because it has a __init__ constructor (from: tests/test_turbulence_verification.py)
    class TestCaseSize(Enum):
