+++ tests/test_turbulence_verification.py
@@ -36,20 +36,20 @@ TestCaseValues = namedtuple("TestCaseValues", ["index_into_dataset", "observatio
 AmdarTurbDataMock = namedtuple("AmdarTurbDataMock", ["amdar_data_mock", "observational_data_mock"])
 
 
-class TestCaseSize(Enum):
+class ThisTestCaseSize(Enum):
     SMALL = auto()
     LARGE = auto()
 
 
-possible_test_case_sizes: list[TestCaseSize] = [TestCaseSize.SMALL, TestCaseSize.LARGE]
+possible_test_case_sizes: list[ThisTestCaseSize] = [ThisTestCaseSize.SMALL, ThisTestCaseSize.LARGE]
 
 
 @pytest.fixture
-def get_test_case(load_cat_data) -> Callable[[TestCaseSize], TestCaseValues]:
-    def _get_test_case(size: TestCaseSize) -> TestCaseValues:
+def get_test_case(load_cat_data) -> Callable[[ThisTestCaseSize], TestCaseValues]:
+    def _get_test_case(size: ThisTestCaseSize) -> TestCaseValues:
         rand_generator = np.random.default_rng()
         match size:
-            case TestCaseSize.SMALL:
+            case ThisTestCaseSize.SMALL:
                 index_into_dataset = [
                     {"latitude": 0, "longitude": 2, "pressure_level": 2, "time": 0},
                     {"latitude": 1, "longitude": 5, "pressure_level": 1, "time": 0},
@@ -67,7 +67,7 @@ def get_test_case(load_cat_data) -> Callable[[TestCaseSize], TestCaseValues]:
                 }
                 observational_data: dd.DataFrame = dd.from_pandas(pd.DataFrame(indices), npartitions=2)
                 return TestCaseValues(index_into_dataset, observational_data, indices)
-            case TestCaseSize.LARGE:
+            case ThisTestCaseSize.LARGE:
                 cat_data: CATData = load_cat_data(None, with_chunks=True)
                 num_points: int = 30
                 indices: dict[str, np.ndarray] = {
@@ -103,7 +103,7 @@ def get_test_case(load_cat_data) -> Callable[[TestCaseSize], TestCaseValues]:
 
 
 @pytest.mark.parametrize("case_size", possible_test_case_sizes)
-def test_observation_coordinates_as_arrays(get_test_case, case_size: TestCaseSize) -> None:
+def test_observation_coordinates_as_arrays(get_test_case, case_size: ThisTestCaseSize) -> None:
     case = get_test_case(case_size)
     coords: ObservationCoordinates = ObservationCoordinates(
         case.observational_data["level_index"].to_dask_array(lengths=True),
@@ -146,7 +146,7 @@ class TestAmdarDataHarmoniser:
     def test_observations_index_to_grid_fails(
         self,
         mocker: "MockerFixture",
-        case_size: TestCaseSize,
+        case_size: ThisTestCaseSize,
         indexing_format: IndexingFormat,
         load_cat_data,
         get_test_case,
@@ -182,7 +182,13 @@ class TestAmdarDataHarmoniser:
     @pytest.mark.parametrize("case_size", possible_test_case_sizes)
     @pytest.mark.parametrize("stack_on_axis", [0, 1])
     def test_observations_index_to_grid_coords(
-        self, mocker: "MockerFixture", stack_on_axis: int, case_size: TestCaseSize, load_cat_data, get_test_case, client
+        self,
+        mocker: "MockerFixture",
+        stack_on_axis: int,
+        case_size: ThisTestCaseSize,
+        load_cat_data,
+        get_test_case,
+        client,
     ) -> None:
         case: TestCaseValues = get_test_case(case_size)
[…]
