### 2026-01-15 9eda540
refactor: Rename class to eliminate PytestCollectionWarning
PytestCollectionWarning: cannot collect test class 'TestCaseValues' because it has a __new__ constructor (from: tests/test_turbulence_verification.py)
