### 2026-01-16 83ade60
refactor: Use as_dataset() method and pass kwargs in more safely
