### 2026-01-16 ca425bc
deps: Relaxed constraint on zarr as it is not necessary
