### 2026-01-16 1515779
refactor: Force matthews_corr_coeff to only accept kwargs so errors are picked up by type checker
