### 2026-01-16 08578a7
refactor: Add output dir to EvaluationStage so I can output stuff there
