### 2026-01-16 f55b03e
feat: Added turbulence analysis class that will compute turbulent regions based on threhsold to return boolean arrrays
