### 2026-01-16 be7a942
feat: Implemented classes to compute the matthews correlation coefficient on a given dataset, and another to coordinate converting turbulence diagnostics to booleans so correaltion can be computed
