### 2026-01-16 a5c790d
maint: Deleted dead code
