### 2026-01-16 5e9197b
fix: Convert denominator to float to resolve np.sqrt error about acting on integer
