### 2026-01-16 b39e83f
feat: Plumbing for matthews correlation coefficient evaluation phase
