### 2026-01-20 57a7709
experiment: swap around netcdf dependencies to see if it fixes issue
