+++ pyproject.toml
@@ -26,7 +26,7 @@ dependencies = [
     "crick", # necessary for dask.percentile method
     "dask",
     "distributed",
-    "h5netcdf",
+    "netcdf4",
     "pydantic",
     "pyproj",
     "pyyaml",
@@ -51,7 +51,7 @@ rojak = "rojak.cli.main:app"
 amdar = [
     "dask-geopandas",
     "geopandas",
-    "netcdf4",
+    "h5netcdf",
     "pyarrow",
 ]
 ecmwf = [
