+++ src/rojak/orchestrator/turbulence.py
@@ -385,6 +385,18 @@ class EvaluationStage:
                 _ = matthews_correlation.to_zarr(
                     self._output_dir / f"matthews_corr_{chained_names}.zarr", mode="w", zarr_format=2
                 )
+                for level in self._config.pressure_levels:
+                    matthews_correlation_on_level: xr.DataArray = MatthewsCorrelationOnThresholdedDiagnostics(
+                        suite.as_dataset(),
+                        self._config.severities,
+                        thresholds,
+                        self._config.threshold_mode,
+                        pressure_level=level,
+                    ).execute()
+                    _ = matthews_correlation_on_level.to_zarr(
+                        self._output_dir / f"matthews_corr_{level:.0f}_{chained_names}.zarr", mode="w", zarr_format=2
+                    )
+
                 return Result(matthews_correlation)
             case _ as unreachable:
                 assert_never(unreachable)
