### 2026-01-22 5594700
feat: Rename dataarray to name of diagnostic
