### 2026-01-22 0516964
fix: Prevent reciprocal from being applied to values=0
