### 2026-01-22 93c3d28
fix: Employ F2D method to prevent 1/0 in F3D
