### 2026-01-22 eca2279
fix: Add missing negative sign
