### 2026-01-22 59fdad6
feat: Replace where with clip
