### 2026-01-22 2378372
feat: Replace where with modulo ufunc
