### 2026-01-22 e718222
fix: Replaced where with clip
