### 2026-01-22 e3b428d
fix: Replaced where with clip
