### 2026-01-22 0e6e8d1
fix: Restore original implementation
