### 2026-01-22 79e619e
fix: Added missing type information for generics
