### 2026-01-22 9bc907e
feat: Added lite application
This is to run rojak in a more light weight manner. The code paths will be more limited. But the goal is to have a set of functionality that works in a lower memory environment
