+++ src/rojak/cli/main.py
@@ -21,12 +21,13 @@ import typer
 from dask.distributed import Client
 from rich.logging import RichHandler
 
-from rojak.cli import data_interface
+from rojak.cli import data_interface, lite_interface
 from rojak.orchestrator.configuration import Context as ConfigContext
 from rojak.orchestrator.turbulence import TurbulenceLauncher
 
 app = typer.Typer()
 app.add_typer(data_interface.data_app, name="data")
+app.add_typer(lite_interface.lite_app, name="lite")
 
 
 class LogLevel(StrEnum):
