### 2026-01-23 65d4894
feat: Initial plumbing to compute distributions parameters throught lite interface
