### 2026-01-23 73fbb37
feat: Implemented interface to compute distribution parameters
