### 2026-01-29 ceadb4e
refactor: generalised type adaptor functionality
Made the exporting to json for the the histogram data, distribution parameters and therholds into a generic function that's implemented by the lite controller. This is now imported by the turbulence orchestrator
