+++ tests/test_orchestrator_turbulence.py
@@ -12,7 +12,8 @@ from rojak.orchestrator.configuration import (
     TurbulenceDiagnostics,
     TurbulenceThresholds,
 )
-from rojak.orchestrator.turbulence import DISTRIBUTION_PARAMS_TYPE_ADAPTER, THRESHOLDS_TYPE_ADAPTER, CalibrationStage
+from rojak.orchestrator.lite_controller import HISTOGRAM_DATA_TYPE_ADAPTER, THRESHOLDS_TYPE_ADAPTER
+from rojak.orchestrator.turbulence import CalibrationStage
 from rojak.turbulence.analysis import HistogramData
 
 if TYPE_CHECKING:
@@ -238,5 +239,5 @@ def test_calibration_stage_compute_distribution_params(
     )
     assert generated_dist_params_file.exists()
     assert generated_dist_params_file.is_file()
-    instantiated_from_generated = DISTRIBUTION_PARAMS_TYPE_ADAPTER.validate_json(generated_dist_params_file.read_text())
+    instantiated_from_generated = HISTOGRAM_DATA_TYPE_ADAPTER.validate_json(generated_dist_params_file.read_text())
     assert instantiated_from_generated == output_dist_params
