### 2026-01-29 981c92d
refactor: Replaced implementation with more efficient ones
