### 2026-01-30 620fb07
refactor: Moved earth average radius into core constants as it will be used somewhere else
