### 2026-01-30 7da54a4
feat: Implemented method to compute haversine distance
