### 2026-01-30 65942bd
feat: implemented method to compute great circle distance from one region to another
This uses the haversine distance to convert degrees of latitude and longitude. It also determines the shortest distance based on the euclidean distance. In theory, this could be swapped out but this features isn't necessary atm. YAGNI
