### 2026-01-31 2064c6f
test: Fix broken test after removing the units control
