### 2026-02-07 33b4f21
feat: Implemented method to compute the shortest vertical distance to another region
This is to understand how far one needs to divert in the vertical direction to hit a CAT region from an ISSR
